Split deserializer tests into several files to speed up compilation
Tests for sequences occupies over 4000 lines of code.
Tests for enums occupies about 1000 lines of code.

Since these tests are loosely related, it makes sense to separate
them into separate files to speed up compilation
