+++ tests/serde-de.rs
@@ -8,27 +8,8 @@ use serde::de::IgnoredAny;
 use serde::serde_if_integer128;
 use serde::Deserialize;
 
-/// Deserialize an instance of type T from a string of XML text.
-/// If deserialization was succeeded checks that all XML events was consumed
-fn from_str<'de, T>(s: &'de str) -> Result<T, DeError>
-where
-    T: Deserialize<'de>,
-{
-    // Log XML that we try to deserialize to see it in the failed tests output
-    dbg!(s);
-    let mut de = Deserializer::from_str(s);
-    let result = T::deserialize(&mut de);
-
-    // If type was deserialized, the whole XML document should be consumed
-    if let Ok(_) = result {
-        match <()>::deserialize(&mut de) {
-            Err(DeError::UnexpectedEof) => (),
-            e => panic!("Expected end `UnexpectedEof`, but got {:?}", e),
-        }
-    }
-
-    result
-}
+mod helpers;
+use helpers::from_str;
 
 /// Tests for deserializing into specially named field `$text` which represent
 /// textual content of an XML element
@@ -471,4085 +452,7 @@ mod tuple_struct {
     }
 }
 
-mod seq {
-    use super::*;
-
-    /// Check that top-level sequences can be deserialized from the multi-root XML documents
-    mod top_level {
-        use super::*;
-        use pretty_assertions::assert_eq;
-
-        #[test]
-        fn simple() {
-            from_str::<[(); 3]>("<root/><root>42</root><root>answer</root>").unwrap();
-
-            let data: Vec<()> = from_str("<root/><root>42</root><root>answer</root>").unwrap();
-            assert_eq!(data, vec![(), (), ()]);
-        }
-
-        /// Special case: empty sequence
-        #[test]
-        fn empty() {
-            from_str::<[(); 0]>("").unwrap();
-
-            let data: Vec<()> = from_str("").unwrap();
-            assert_eq!(data, vec![]);
-        }
-
-        /// Special case: one-element sequence
-        #[test]
-        fn one_element() {
-            from_str::<[(); 1]>("<root/>").unwrap();
-            from_str::<[(); 1]>("<root>42</root>").unwrap();
-            from_str::<[(); 1]>("text").unwrap();
-            from_str::<[(); 1]>("<![CDATA[cdata]]>").unwrap();
-
-            let data: Vec<()> = from_str("<root/>").unwrap();
-            assert_eq!(data, vec![()]);
-
-            let data: Vec<()> = from_str("<root>42</root>").unwrap();
-            assert_eq!(data, vec![()]);
-
-            let data: Vec<()> = from_str("text").unwrap();
-            assert_eq!(data, vec![()]);
-
-            let data: Vec<()> = from_str("<![CDATA[cdata]]>").unwrap();
-            assert_eq!(data, vec![()]);
-        }
-
-        #[test]
-        fn excess_attribute() {
-            from_str::<[(); 3]>(r#"<root/><root excess="attribute">42</root><root>answer</root>"#)
-                .unwrap();
-
-            let data: Vec<()> =
-                from_str(r#"<root/><root excess="attribute">42</root><root>answer</root>"#)
-                    .unwrap();
-            assert_eq!(data, vec![(), (), ()]);
-        }
-
-        #[test]
-        fn mixed_content() {
-            // Text and CDATA represents a one logical text item
-            from_str::<[(); 2]>(
-                r#"
-                <element/>
-                text
-                <![CDATA[cdata]]>
-                "#,
-            )
-            .unwrap();
-
-            let data: Vec<()> = from_str(
-                r#"
-                <element/>
-                text
-                <![CDATA[cdata]]>
-                "#,
-            )
-            .unwrap();
-            // Text and CDATA represents a one logical text item
-            assert_eq!(data, vec![(), ()]);
-        }
-
-        /// This test ensures that composition of deserializer building blocks plays well
-        #[test]
-        fn list_of_struct() {
-            #[derive(Debug, PartialEq, Default, Deserialize)]
-            #[serde(default)]
-            struct Struct {
-                #[serde(rename = "@attribute")]
-                attribute: Option<String>,
-                element: Option<String>,
-            }
-
-            let data: Vec<Struct> = from_str(
-                r#"
-                <struct/>
-                <struct attribute="value"/>
-                <struct>
-                    <element>value</element>
-                </struct>
-                <struct attribute="value">
-                    <element>value</element>
-                </struct>
-                "#,
-            )
-            .unwrap();
-            assert_eq!(
-                data,
-                vec![
-                    Struct {
-                        attribute: None,
-                        element: None,
-                    },
-                    Struct {
-                        attribute: Some("value".to_string()),
-                        element: None,
-                    },
-                    Struct {
-                        attribute: None,
-                        element: Some("value".to_string()),
-                    },
-                    Struct {
-                        attribute: Some("value".to_string()),
-                        element: Some("value".to_string()),
-                    },
-                ]
-            );
-        }
-
-        /// Test for https://github.com/tafia/quick-xml/issues/500
-        #[test]
-        fn list_of_enum() {
-            #[derive(Debug, PartialEq, Deserialize)]
-            enum Enum {
-                One,
-                Two,
-            }
-
-            let data: Vec<Enum> = from_str(
-                r#"
-                <One/>
-                <Two/>
-                <One/>
-                "#,
-            )
-            .unwrap();
-            assert_eq!(data, vec![Enum::One, Enum::Two, Enum::One]);
-        }
-    }
-
-    /// Tests where each sequence item have an identical name in an XML.
-    /// That explicitly means that `enum`s as list elements are not supported
-    /// in that case, because enum requires different tags.
-    ///
-    /// (by `enums` we mean [externally tagged enums] is serde terminology)
-    ///
-    /// [externally tagged enums]: https://serde.rs/enum-representations.html#externally-tagged
-    mod fixed_name {
-        use super::*;
-
-        /// This module contains tests where size of the list have a compile-time size
-        mod fixed_size {
-            use super::*;
-            use pretty_assertions::assert_eq;
-
-            #[derive(Debug, PartialEq, Deserialize)]
-            struct List {
-                item: [(); 3],
-            }
-
-            /// Simple case: count of elements matches expected size of sequence,
-            /// each element has the same name. Successful deserialization expected
-            #[test]
-            fn simple() {
-                from_str::<List>(
-                    r#"
-                    <root>
-                        <item/>
-                        <item/>
-                        <item/>
-                    </root>
-                    "#,
-                )
-                .unwrap();
-            }
-
-            /// Special case: empty sequence
-            #[test]
-            #[ignore = "it is impossible to distinguish between missed field and empty list: use `Option<>` or #[serde(default)]"]
-            fn empty() {
-                #[derive(Debug, PartialEq, Deserialize)]
-                struct List {
-                    item: [(); 0],
-                }
-
-                from_str::<List>(r#"<root></root>"#).unwrap();
-                from_str::<List>(r#"<root/>"#).unwrap();
-            }
-
-            /// Special case: one-element sequence
-            #[test]
-            fn one_element() {
-                #[derive(Debug, PartialEq, Deserialize)]
-                struct List {
-                    item: [(); 1],
-                }
-
-                from_str::<List>(
-                    r#"
-                    <root>
-                        <item/>
-                    </root>
-                    "#,
-                )
-                .unwrap();
-            }
-
-            /// Fever elements than expected size of sequence, each element has
-            /// the same name. Failure expected
-            #[test]
-            fn fever_elements() {
-                let data = from_str::<List>(
-                    r#"
-                    <root>
-                        <item/>
-                        <item/>
-                    </root>
-                    "#,
-                );
-
-                match data {
-                    Err(DeError::Custom(e)) => {
-                        assert_eq!(e, "invalid length 2, expected an array of length 3")
-                    }
-                    e => panic!(
-                        r#"Expected `Err(Custom("invalid length 2, expected an array of length 3"))`, but found {:?}"#,
-                        e
-                    ),
-                }
-            }
-
-            /// More elements than expected size of sequence, each element has
-            /// the same name. Failure expected. If you wish to ignore excess
-            /// elements, use the special type, that consume as much elements
-            /// as possible, but ignores excess elements
-            #[test]
-            fn excess_elements() {
-                let data = from_str::<List>(
-                    r#"
-                    <root>
-                        <item/>
-                        <item/>
-                        <item/>
-                        <item/>
-                    </root>
-                    "#,
-                );
-
-                match data {
-                    Err(DeError::Custom(e)) => assert_eq!(e, "duplicate field `item`"),
-                    e => panic!(
-                        r#"Expected `Err(Custom("duplicate field `item`"))`, but found {:?}"#,
-                        e
-                    ),
-                }
-            }
-
-            /// Mixed content assumes, that some elements will have an internal
-            /// name `$text` or `$value`, so, unless field named the same, it is expected
-            /// to fail
-            #[test]
-            fn mixed_content() {
-                let data = from_str::<List>(
-                    r#"
-                    <root>
-                        <element/>
-                        text
-                        <![CDATA[cdata]]>
-                    </root>
-                    "#,
-                );
-
-                match data {
-                    Err(DeError::Custom(e)) => assert_eq!(e, "missing field `item`"),
-                    e => panic!(
-                        r#"Expected `Err(Custom("missing field `item`"))`, but found {:?}"#,
-                        e
-                    ),
-                }
-            }
-
-            /// In those tests sequence should be deserialized from an XML
-            /// with additional elements that is not defined in the struct.
-            /// That fields should be skipped during deserialization
-            mod unknown_items {
-                use super::*;
-                #[cfg(not(feature = "overlapped-lists"))]
-                use pretty_assertions::assert_eq;
-
-                #[test]
-                fn before() {
-                    from_str::<List>(
-                        r#"
-                        <root>
-                            <unknown/>
-                            <item/>
-                            <item/>
-                            <item/>
-                        </root>
-                        "#,
-                    )
-                    .unwrap();
-                }
-
-                #[test]
-                fn after() {
-                    from_str::<List>(
-                        r#"
-                        <root>
-                            <item/>
-                            <item/>
-                            <item/>
-                            <unknown/>
-                        </root>
-                        "#,
-                    )
-                    .unwrap();
-                }
-
-                #[test]
-                fn overlapped() {
-                    let data = from_str::<List>(
-                        r#"
-                        <root>
-                            <item/>
-                            <unknown/>
-                            <item/>
-                            <item/>
-                        </root>
-                        "#,
-                    );
-
-                    #[cfg(feature = "overlapped-lists")]
-                    data.unwrap();
-
-                    #[cfg(not(feature = "overlapped-lists"))]
-                    match data {
-                        Err(DeError::Custom(e)) => {
-                            assert_eq!(e, "invalid length 1, expected an array of length 3")
-                        }
-                        e => panic!(
-                            r#"Expected Err(Custom("invalid length 1, expected an array of length 3")), got {:?}"#,
-                            e
-                        ),
-                    }
-                }
-
-                /// Test for https://github.com/tafia/quick-xml/issues/435
-                #[test]
-                fn overlapped_with_nested_list() {
-                    #[derive(Debug, PartialEq, Deserialize)]
-                    struct Root {
-                        outer: [List; 3],
-                    }
-
-                    let data = from_str::<Root>(
-                        r#"
-                        <root>
-                          <outer><item/><item/><item/></outer>
-                          <unknown/>
-                          <outer><item/><item/><item/></outer>
-                          <outer><item/><item/><item/></outer>
-                        </root>
-                        "#,
-                    );
-
-                    #[cfg(feature = "overlapped-lists")]
-                    data.unwrap();
-
-                    #[cfg(not(feature = "overlapped-lists"))]
-                    match data {
-                        Err(DeError::Custom(e)) => {
-                            assert_eq!(e, "invalid length 1, expected an array of length 3")
-                        }
-                        e => panic!(
-                            r#"Expected Err(Custom("invalid length 1, expected an array of length 3")), got {:?}"#,
-                            e
-                        ),
-                    }
-                }
-            }
-
-            /// In those tests non-sequential field is defined in the struct
-            /// before sequential, so it will be deserialized before the list.
-            /// That struct should be deserialized from an XML where these
-            /// fields comes in an arbitrary order
-            mod field_before_list {
-                use super::*;
-                #[cfg(not(feature = "overlapped-lists"))]
-                use pretty_assertions::assert_eq;
-
-                #[derive(Debug, PartialEq, Deserialize)]
-                struct Root {
-                    node: (),
-                    item: [(); 3],
-                }
-
-                #[test]
-                fn before() {
-                    from_str::<Root>(
-                        r#"
-                        <root>
-                            <node/>
-                            <item/>
-                            <item/>
-                            <item/>
-                        </root>
-                        "#,
-                    )
-                    .unwrap();
-                }
-
-                #[test]
-                fn after() {
-                    from_str::<Root>(
-                        r#"
-                        <root>
-                            <item/>
-                            <item/>
-                            <item/>
-                            <node/>
-                        </root>
-                        "#,
-                    )
-                    .unwrap();
-                }
-
-                #[test]
-                fn overlapped() {
-                    let data = from_str::<Root>(
-                        r#"
-                        <root>
-                            <item/>
-                            <node/>
-                            <item/>
-                            <item/>
-                        </root>
-                        "#,
-                    );
-
-                    #[cfg(feature = "overlapped-lists")]
-                    data.unwrap();
-
-                    #[cfg(not(feature = "overlapped-lists"))]
-                    match data {
-                        Err(DeError::Custom(e)) => {
-                            assert_eq!(e, "invalid length 1, expected an array of length 3")
-                        }
-                        e => panic!(
-                            r#"Expected Err(Custom("invalid length 1, expected an array of length 3")), got {:?}"#,
-                            e
-                        ),
-                    }
-                }
-
-                /// Test for https://github.com/tafia/quick-xml/issues/435
-                #[test]
-                fn overlapped_with_nested_list() {
-                    #[derive(Debug, PartialEq, Deserialize)]
-                    struct Root {
-                        node: (),
-                        outer: [List; 3],
-                    }
-
-                    let data = from_str::<Root>(
-                        r#"
-                        <root>
-                            <outer><item/><item/><item/></outer>
-                            <node/>
-                            <outer><item/><item/><item/></outer>
-                            <outer><item/><item/><item/></outer>
-                        </root>
-                        "#,
-                    );
-
-                    #[cfg(feature = "overlapped-lists")]
-                    data.unwrap();
-
-                    #[cfg(not(feature = "overlapped-lists"))]
-                    match data {
-                        Err(DeError::Custom(e)) => {
-                            assert_eq!(e, "invalid length 1, expected an array of length 3")
-                        }
-                        e => panic!(
-                            r#"Expected Err(Custom("invalid length 1, expected an array of length 3")), got {:?}"#,
-                            e
-                        ),
-                    }
-                }
-            }
-
-            /// In those tests non-sequential field is defined in the struct
-            /// after sequential, so it will be deserialized after the list.
-            /// That struct should be deserialized from an XML where these
-            /// fields comes in an arbitrary order
-            mod field_after_list {
-                use super::*;
-                #[cfg(not(feature = "overlapped-lists"))]
-                use pretty_assertions::assert_eq;
-
-                #[derive(Debug, PartialEq, Deserialize)]
-                struct Root {
-                    item: [(); 3],
-                    node: (),
-                }
-
-                #[test]
-                fn before() {
-                    from_str::<Root>(
-                        r#"
-                        <root>
-                            <node/>
-                            <item/>
-                            <item/>
-                            <item/>
-                        </root>
-                        "#,
-                    )
-                    .unwrap();
-                }
-
-                #[test]
-                fn after() {
-                    from_str::<Root>(
-                        r#"
-                        <root>
-                            <item/>
-                            <item/>
-                            <item/>
-                            <node/>
-                        </root>
-                        "#,
-                    )
-                    .unwrap();
-                }
-
-                #[test]
-                fn overlapped() {
-                    let data = from_str::<Root>(
-                        r#"
-                        <root>
-                            <item/>
-                            <node/>
-                            <item/>
-                            <item/>
-                        </root>
-                        "#,
-                    );
-
-                    #[cfg(feature = "overlapped-lists")]
-                    data.unwrap();
-
-                    #[cfg(not(feature = "overlapped-lists"))]
-                    match data {
-                        Err(DeError::Custom(e)) => {
-                            assert_eq!(e, "invalid length 1, expected an array of length 3")
-                        }
-                        e => panic!(
-                            r#"Expected Err(Custom("invalid length 1, expected an array of length 3")), got {:?}"#,
-                            e
-                        ),
-                    }
-                }
-
-                /// Test for https://github.com/tafia/quick-xml/issues/435
-                #[test]
-                fn overlapped_with_nested_list() {
-                    #[derive(Debug, PartialEq, Deserialize)]
-                    struct Root {
-                        outer: [List; 3],
-                        node: (),
-                    }
-
-                    let data = from_str::<Root>(
-                        r#"
-                        <root>
-                            <outer><item/><item/><item/></outer>
-                            <node/>
-                            <outer><item/><item/><item/></outer>
-                            <outer><item/><item/><item/></outer>
-                        </root>
-                        "#,
-                    );
-
-                    #[cfg(feature = "overlapped-lists")]
-                    data.unwrap();
-
-                    #[cfg(not(feature = "overlapped-lists"))]
-                    match data {
-                        Err(DeError::Custom(e)) => {
-                            assert_eq!(e, "invalid length 1, expected an array of length 3")
-                        }
-                        e => panic!(
-                            r#"Expected Err(Custom("invalid length 1, expected an array of length 3")), got {:?}"#,
-                            e
-                        ),
-                    }
-                }
-            }
-
-            /// In those tests two lists are deserialized simultaneously.
-            /// Lists should be deserialized even when them overlaps
-            mod two_lists {
-                use super::*;
-                #[cfg(not(feature = "overlapped-lists"))]
-                use pretty_assertions::assert_eq;
-
-                #[derive(Debug, PartialEq, Deserialize)]
-                struct Pair {
-                    item: [(); 3],
-                    element: [(); 2],
-                }
-
-                #[test]
-                fn splitted() {
-                    from_str::<Pair>(
-                        r#"
-                        <root>
-                            <element/>
-                            <element/>
-                            <item/>
-                            <item/>
-                            <item/>
-                        </root>
-                        "#,
-                    )
-                    .unwrap();
-                }
-
-                #[test]
-                fn overlapped() {
-                    let data = from_str::<Pair>(
-                        r#"
-                        <root>
-                            <item/>
-                            <element/>
-                            <item/>
-                            <element/>
-                            <item/>
-                        </root>
-                        "#,
-                    );
-
-                    #[cfg(feature = "overlapped-lists")]
-                    data.unwrap();
-
-                    #[cfg(not(feature = "overlapped-lists"))]
-                    match data {
-                        Err(DeError::Custom(e)) => {
-                            assert_eq!(e, "invalid length 1, expected an array of length 3")
-                        }
-                        e => panic!(
-                            r#"Expected Err(Custom("invalid length 1, expected an array of length 3")), got {:?}"#,
-                            e
-                        ),
-                    }
-                }
-
-                /// Test for https://github.com/tafia/quick-xml/issues/435
-                #[test]
-                fn overlapped_with_nested_list() {
-                    #[derive(Debug, PartialEq, Deserialize)]
-                    struct Pair {
-                        outer: [List; 3],
-                        element: [(); 2],
-                    }
-
-                    let data = from_str::<Pair>(
-                        r#"
-                        <root>
-                            <outer><item/><item/><item/></outer>
-                            <element/>
-                            <outer><item/><item/><item/></outer>
-                            <element/>
-                            <outer><item/><item/><item/></outer>
-                        </root>
-                        "#,
-                    );
-
-                    #[cfg(feature = "overlapped-lists")]
-                    data.unwrap();
-
-                    #[cfg(not(feature = "overlapped-lists"))]
-                    match data {
-                        Err(DeError::Custom(e)) => {
-                            assert_eq!(e, "invalid length 1, expected an array of length 3")
-                        }
-                        e => panic!(
-                            r#"Expected Err(Custom("invalid length 1, expected an array of length 3")), got {:?}"#,
-                            e
-                        ),
-                    }
-                }
-            }
-
-            /// Deserialization of primitives slightly differs from deserialization
-            /// of complex types, so need to check this separately
-            #[test]
-            fn primitives() {
-                #[derive(Debug, PartialEq, Deserialize)]
-                struct List {
-                    item: [usize; 3],
-                }
-
-                let data: List = from_str(
-                    r#"
-                    <root>
-                        <item>41</item>
-                        <item>42</item>
-                        <item>43</item>
-                    </root>
-                    "#,
-                )
-                .unwrap();
-                assert_eq!(data, List { item: [41, 42, 43] });
-
-                from_str::<List>(
-                    r#"
-                    <root>
-                        <item>41</item>
-                        <item><item>42</item></item>
-                        <item>43</item>
-                    </root>
-                    "#,
-                )
-                .unwrap_err();
-            }
-
-            /// This test ensures that composition of deserializer building blocks
-            /// plays well
-            #[test]
-            fn list_of_struct() {
-                #[derive(Debug, PartialEq, Default, Deserialize)]
-                #[serde(default)]
-                struct Struct {
-                    #[serde(rename = "@attribute")]
-                    attribute: Option<String>,
-                    element: Option<String>,
-                }
-
-                #[derive(Debug, PartialEq, Deserialize)]
-                struct List {
-                    item: [Struct; 4],
-                }
-
-                let data: List = from_str(
-                    r#"
-                    <root>
-                        <item/>
-                        <item attribute="value"/>
-                        <item>
-                            <element>value</element>
-                        </item>
-                        <item attribute="value">
-                            <element>value</element>
-                        </item>
-                    </root>
-                    "#,
-                )
-                .unwrap();
-                assert_eq!(
-                    data,
-                    List {
-                        item: [
-                            Struct {
-                                attribute: None,
-                                element: None,
-                            },
-                            Struct {
-                                attribute: Some("value".to_string()),
-                                element: None,
-                            },
-                            Struct {
-                                attribute: None,
-                                element: Some("value".to_string()),
-                            },
-                            Struct {
-                                attribute: Some("value".to_string()),
-                                element: Some("value".to_string()),
-                            },
-                        ],
-                    }
-                );
-            }
-
-            /// Checks that sequences represented by elements can contain sequences,
-            /// represented by [`xs:list`s](https://www.w3schools.com/xml/el_list.asp)
-            mod xs_list {
-                use super::*;
-                use pretty_assertions::assert_eq;
-
-                /// Special case: zero elements
-                #[test]
-                fn zero() {
-                    #[derive(Debug, Deserialize, PartialEq)]
-                    struct List {
-                        /// Outer list mapped to elements, inner -- to `xs:list`.
-                        ///
-                        /// `#[serde(default)]` is required to correctly deserialize
-                        /// empty sequence, because without elements the field
-                        /// also is missing and derived `Deserialize` implementation
-                        /// would complain about that unless field is marked as
-                        /// `default`.
-                        #[serde(default)]
-                        item: [Vec<String>; 0],
-                    }
-
-                    let data: List = from_str(
-                        r#"
-                        <root>
-                        </root>
-                        "#,
-                    )
-                    .unwrap();
-
-                    assert_eq!(data, List { item: [] });
-                }
-
-                /// Special case: one element
-                #[test]
-                fn one() {
-                    #[derive(Debug, Deserialize, PartialEq)]
-                    struct List {
-                        /// Outer list mapped to elements, inner -- to `xs:list`.
-                        ///
-                        /// `#[serde(default)]` is not required, because correct
-                        /// XML will always contains at least 1 element.
-                        item: [Vec<String>; 1],
-                    }
-
-                    let data: List = from_str(
-                        r#"
-                        <root>
-                            <item>first list</item>
-                        </root>
-                        "#,
-                    )
-                    .unwrap();
-
-                    assert_eq!(
-                        data,
-                        List {
-                            item: [vec!["first".to_string(), "list".to_string()]]
-                        }
-                    );
-                }
-
-                /// Special case: outer list is always mapped to an elements sequence,
-                /// not to an `xs:list`
-                #[test]
-                fn element() {
-                    #[derive(Debug, Deserialize, PartialEq)]
-                    struct List {
-                        /// Outer list mapped to elements, inner -- to `xs:list`.
-                        ///
-                        /// `#[serde(default)]` is not required, because correct
-                        /// XML will always contains at least 1 element.
-                        item: [String; 1],
-                    }
-
-                    let data: List = from_str(
-                        r#"
-                        <root>
-                            <item>first item</item>
-                        </root>
-                        "#,
-                    )
-                    .unwrap();
-
-                    assert_eq!(
-                        data,
-                        List {
-                            item: ["first item".to_string()]
-                        }
-                    );
-                }
-
-                /// This tests demonstrates, that for `$value` field (`list`) actual
-                /// name of XML element (`item`) does not matter. That allows list
-                /// item to be an enum, where tag name determines enum variant
-                #[test]
-                fn many() {
-                    #[derive(Debug, Deserialize, PartialEq)]
-                    struct List {
-                        /// Outer list mapped to elements, inner -- to `xs:list`.
-                        ///
-                        /// `#[serde(default)]` is not required, because correct
-                        /// XML will always contains at least 1 element.
-                        item: [Vec<String>; 2],
-                    }
-
-                    let data: List = from_str(
-                        r#"
-                        <root>
-                            <item>first list</item>
-                            <item>second list</item>
-                        </root>
-                        "#,
-                    )
-                    .unwrap();
-
-                    assert_eq!(
-                        data,
-                        List {
-                            item: [
-                                vec!["first".to_string(), "list".to_string()],
-                                vec!["second".to_string(), "list".to_string()],
-                            ]
-                        }
-                    );
-                }
-            }
-        }
-
-        /// This module contains tests where size of the list have an unspecified size
-        mod variable_size {
-            use super::*;
-            use pretty_assertions::assert_eq;
-
-            #[derive(Debug, PartialEq, Deserialize)]
-            struct List {
-                item: Vec<()>,
-            }
-
-            /// Simple case: count of elements matches expected size of sequence,
-            /// each element has the same name. Successful deserialization expected
-            #[test]
-            fn simple() {
-                let data: List = from_str(
-                    r#"
-                    <root>
-                        <item/>
-                        <item/>
-                        <item/>
-                    </root>
-                    "#,
-                )
-                .unwrap();
-
-                assert_eq!(
-                    data,
-                    List {
-                        item: vec![(), (), ()],
-                    }
-                );
-            }
-
-            /// Special case: empty sequence
-            #[test]
-            #[ignore = "it is impossible to distinguish between missed field and empty list: use `Option<>` or #[serde(default)]"]
-            fn empty() {
-                let data: List = from_str(r#"<root></root>"#).unwrap();
-                assert_eq!(data, List { item: vec![] });
-
-                let data: List = from_str(r#"<root/>"#).unwrap();
-                assert_eq!(data, List { item: vec![] });
-            }
-
-            /// Special case: one-element sequence
-            #[test]
-            fn one_element() {
-                let data: List = from_str(
-                    r#"
-                    <root>
-                        <item/>
-                    </root>
-                    "#,
-                )
-                .unwrap();
-
-                assert_eq!(data, List { item: vec![()] });
-            }
-
-            /// Mixed content assumes, that some elements will have an internal
-            /// name `$text` or `$value`, so, unless field named the same, it is expected
-            /// to fail
-            #[test]
-            fn mixed_content() {
-                let data = from_str::<List>(
-                    r#"
-                    <root>
-                        <element/>
-                        text
-                        <![CDATA[cdata]]>
-                    </root>
-                    "#,
-                );
-
-                match data {
-                    Err(DeError::Custom(e)) => assert_eq!(e, "missing field `item`"),
-                    e => panic!(
-                        r#"Expected `Err(Custom("missing field `item`"))`, but found {:?}"#,
-                        e
-                    ),
-                }
-            }
-
-            /// In those tests sequence should be deserialized from the XML
-            /// with additional elements that is not defined in the struct.
-            /// That fields should be skipped during deserialization
-            mod unknown_items {
-                use super::*;
-                use pretty_assertions::assert_eq;
-
-                #[test]
-                fn before() {
-                    let data: List = from_str(
-                        r#"
-                        <root>
-                            <unknown/>
-                            <item/>
-                            <item/>
-                            <item/>
-                        </root>
-                        "#,
-                    )
-                    .unwrap();
-
-                    assert_eq!(
-                        data,
-                        List {
-                            item: vec![(), (), ()],
-                        }
-                    );
-                }
-
-                #[test]
-                fn after() {
-                    let data: List = from_str(
-                        r#"
-                        <root>
-                            <item/>
-                            <item/>
-                            <item/>
-                            <unknown/>
-                        </root>
-                        "#,
-                    )
-                    .unwrap();
-
-                    assert_eq!(
-                        data,
-                        List {
-                            item: vec![(), (), ()],
-                        }
-                    );
-                }
-
-                #[test]
-                fn overlapped() {
-                    let data = from_str::<List>(
-                        r#"
-                        <root>
-                            <item/>
-                            <unknown/>
-                            <item/>
-                            <item/>
-                        </root>
-                        "#,
-                    );
-
-                    #[cfg(feature = "overlapped-lists")]
-                    assert_eq!(
-                        data.unwrap(),
-                        List {
-                            item: vec![(), (), ()],
-                        }
-                    );
-
-                    #[cfg(not(feature = "overlapped-lists"))]
-                    match data {
-                        Err(DeError::Custom(e)) => assert_eq!(e, "duplicate field `item`"),
-                        e => panic!(
-                            r#"Expected Err(Custom("duplicate field `item`")), got {:?}"#,
-                            e
-                        ),
-                    }
-                }
-
-                /// Test for https://github.com/tafia/quick-xml/issues/435
-                #[test]
-                fn overlapped_with_nested_list() {
-                    #[derive(Debug, PartialEq, Deserialize)]
-                    struct Root {
-                        outer: Vec<List>,
-                    }
-
-                    let data = from_str::<Root>(
-                        r#"
-                        <root>
-                            <outer><item/></outer>
-                            <unknown/>
-                            <outer><item/></outer>
-                            <outer><item/></outer>
-                        </root>
-                        "#,
-                    );
-
-                    #[cfg(feature = "overlapped-lists")]
-                    assert_eq!(
-                        data.unwrap(),
-                        Root {
-                            outer: vec![
-                                List { item: vec![()] },
-                                List { item: vec![()] },
-                                List { item: vec![()] },
-                            ],
-                        }
-                    );
-
-                    #[cfg(not(feature = "overlapped-lists"))]
-                    match data {
-                        Err(DeError::Custom(e)) => assert_eq!(e, "duplicate field `outer`"),
-                        e => panic!(
-                            r#"Expected Err(Custom("duplicate field `outer`")), got {:?}"#,
-                            e
-                        ),
-                    }
-                }
-            }
-
-            /// In those tests non-sequential field is defined in the struct
-            /// before sequential, so it will be deserialized before the list.
-            /// That struct should be deserialized from the XML where these
-            /// fields comes in an arbitrary order
-            mod field_before_list {
-                use super::*;
-                use pretty_assertions::assert_eq;
-
-                #[derive(Debug, PartialEq, Default, Deserialize)]
-                struct Root {
-                    node: (),
-                    item: Vec<()>,
-                }
-
-                #[test]
-                fn before() {
-                    let data: Root = from_str(
-                        r#"
-                        <root>
-                            <node/>
-                            <item/>
-                            <item/>
-                            <item/>
-                        </root>
-                        "#,
-                    )
-                    .unwrap();
-
-                    assert_eq!(
-                        data,
-                        Root {
-                            node: (),
-                            item: vec![(), (), ()],
-                        }
-                    );
-                }
-
-                #[test]
-                fn after() {
-                    let data: Root = from_str(
-                        r#"
-                        <root>
-                            <item/>
-                            <item/>
-                            <item/>
-                            <node/>
-                        </root>
-                        "#,
-                    )
-                    .unwrap();
-
-                    assert_eq!(
-                        data,
-                        Root {
-                            node: (),
-                            item: vec![(), (), ()],
-                        }
-                    );
-                }
-
-                #[test]
-                fn overlapped() {
-                    let data = from_str::<Root>(
-                        r#"
-                        <root>
-                            <item/>
-                            <node/>
-                            <item/>
-                            <item/>
-                        </root>
-                        "#,
-                    );
-
-                    #[cfg(feature = "overlapped-lists")]
-                    assert_eq!(
-                        data.unwrap(),
-                        Root {
-                            node: (),
-                            item: vec![(), (), ()],
-                        }
-                    );
-
-                    #[cfg(not(feature = "overlapped-lists"))]
-                    match data {
-                        Err(DeError::Custom(e)) => {
-                            assert_eq!(e, "duplicate field `item`")
-                        }
-                        e => panic!(
-                            r#"Expected Err(Custom("duplicate field `item`")), got {:?}"#,
-                            e
-                        ),
-                    }
-                }
-
-                /// Test for https://github.com/tafia/quick-xml/issues/435
-                #[test]
-                fn overlapped_with_nested_list() {
-                    #[derive(Debug, PartialEq, Deserialize)]
-                    struct Root {
-                        node: (),
-                        outer: Vec<List>,
-                    }
-
-                    let data = from_str::<Root>(
-                        r#"
-                        <root>
-                            <outer><item/></outer>
-                            <node/>
-                            <outer><item/></outer>
-                            <outer><item/></outer>
-                        </root>
-                        "#,
-                    );
-
-                    #[cfg(feature = "overlapped-lists")]
-                    assert_eq!(
-                        data.unwrap(),
-                        Root {
-                            node: (),
-                            outer: vec![
-                                List { item: vec![()] },
-                                List { item: vec![()] },
-                                List { item: vec![()] },
-                            ],
-                        }
-                    );
-
-                    #[cfg(not(feature = "overlapped-lists"))]
-                    match data {
-                        Err(DeError::Custom(e)) => assert_eq!(e, "duplicate field `outer`"),
-                        e => panic!(
-                            r#"Expected Err(Custom("duplicate field `outer`")), got {:?}"#,
-                            e
-                        ),
-                    }
-                }
-            }
-
-            /// In those tests non-sequential field is defined in the struct
-            /// after sequential, so it will be deserialized after the list.
-            /// That struct should be deserialized from the XML where these
-            /// fields comes in an arbitrary order
-            mod field_after_list {
-                use super::*;
-                use pretty_assertions::assert_eq;
-
-                #[derive(Debug, PartialEq, Default, Deserialize)]
-                struct Root {
-                    item: Vec<()>,
-                    node: (),
-                }
-
-                #[test]
-                fn before() {
-                    let data: Root = from_str(
-                        r#"
-                        <root>
-                            <node/>
-                            <item/>
-                            <item/>
-                            <item/>
-                        </root>
-                        "#,
-                    )
-                    .unwrap();
-
-                    assert_eq!(
-                        data,
-                        Root {
-                            item: vec![(), (), ()],
-                            node: (),
-                        }
-                    );
-                }
-
-                #[test]
-                fn after() {
-                    let data: Root = from_str(
-                        r#"
-                        <root>
-                            <item/>
-                            <item/>
-                            <item/>
-                            <node/>
-                        </root>
-                        "#,
-                    )
-                    .unwrap();
-
-                    assert_eq!(
-                        data,
-                        Root {
-                            item: vec![(), (), ()],
-                            node: (),
-                        }
-                    );
-                }
-
-                #[test]
-                fn overlapped() {
-                    let data = from_str::<Root>(
-                        r#"
-                        <root>
-                            <item/>
-                            <node/>
-                            <item/>
-                            <item/>
-                        </root>
-                        "#,
-                    );
-
-                    #[cfg(feature = "overlapped-lists")]
-                    assert_eq!(
-                        data.unwrap(),
-                        Root {
-                            item: vec![(), (), ()],
-                            node: (),
-                        }
-                    );
-
-                    #[cfg(not(feature = "overlapped-lists"))]
-                    match data {
-                        Err(DeError::Custom(e)) => {
-                            assert_eq!(e, "duplicate field `item`")
-                        }
-                        e => panic!(
-                            r#"Expected Err(Custom("duplicate field `item`")), got {:?}"#,
-                            e
-                        ),
-                    }
-                }
-
-                /// Test for https://github.com/tafia/quick-xml/issues/435
-                #[test]
-                fn overlapped_with_nested_list() {
-                    #[derive(Debug, PartialEq, Deserialize)]
-                    struct Root {
-                        outer: Vec<List>,
-                        node: (),
-                    }
-
-                    let data = from_str::<Root>(
-                        r#"
-                        <root>
-                            <outer><item/></outer>
-                            <node/>
-                            <outer><item/></outer>
-                            <outer><item/></outer>
-                        </root>
-                        "#,
-                    );
-
-                    #[cfg(feature = "overlapped-lists")]
-                    assert_eq!(
-                        data.unwrap(),
-                        Root {
-                            outer: vec![
-                                List { item: vec![()] },
-                                List { item: vec![()] },
-                                List { item: vec![()] },
-                            ],
-                            node: (),
-                        }
-                    );
-
-                    #[cfg(not(feature = "overlapped-lists"))]
-                    match data {
-                        Err(DeError::Custom(e)) => assert_eq!(e, "duplicate field `outer`"),
-                        e => panic!(
-                            r#"Expected Err(Custom("duplicate field `outer`")), got {:?}"#,
-                            e
-                        ),
-                    }
-                }
-            }
-
-            /// In those tests two lists are deserialized simultaneously.
-            /// Lists should be deserialized even when them overlaps
-            mod two_lists {
-                use super::*;
-                use pretty_assertions::assert_eq;
-
-                #[derive(Debug, PartialEq, Deserialize)]
-                struct Pair {
-                    item: Vec<()>,
-                    element: Vec<()>,
-                }
-
-                #[test]
-                fn splitted() {
-                    let data: Pair = from_str(
-                        r#"
-                        <root>
-                            <element/>
-                            <element/>
-                            <item/>
-                            <item/>
-                            <item/>
-                        </root>
-                        "#,
-                    )
-                    .unwrap();
-
-                    assert_eq!(
-                        data,
-                        Pair {
-                            item: vec![(), (), ()],
-                            element: vec![(), ()],
-                        }
-                    );
-                }
-
-                #[test]
-                fn overlapped() {
-                    let data = from_str::<Pair>(
-                        r#"
-                        <root>
-                            <item/>
-                            <element/>
-                            <item/>
-                            <element/>
-                            <item/>
-                        </root>
-                        "#,
-                    );
-
-                    #[cfg(feature = "overlapped-lists")]
-                    assert_eq!(
-                        data.unwrap(),
-                        Pair {
-                            item: vec![(), (), ()],
-                            element: vec![(), ()],
-                        }
-                    );
-
-                    #[cfg(not(feature = "overlapped-lists"))]
-                    match data {
-                        Err(DeError::Custom(e)) => assert_eq!(e, "duplicate field `item`"),
-                        e => panic!(
-                            r#"Expected Err(Custom("duplicate field `item`")), got {:?}"#,
-                            e
-                        ),
-                    }
-                }
-
-                #[test]
-                fn overlapped_with_nested_list() {
-                    #[derive(Debug, PartialEq, Deserialize)]
-                    struct Pair {
-                        outer: Vec<List>,
-                        element: Vec<()>,
-                    }
-
-                    let data = from_str::<Pair>(
-                        r#"
-                        <root>
-                            <outer><item/></outer>
-                            <element/>
-                            <outer><item/></outer>
-                            <outer><item/></outer>
-                        </root>
-                        "#,
-                    );
-
-                    #[cfg(feature = "overlapped-lists")]
-                    assert_eq!(
-                        data.unwrap(),
-                        Pair {
-                            outer: vec![
-                                List { item: vec![()] },
-                                List { item: vec![()] },
-                                List { item: vec![()] },
-                            ],
-                            element: vec![()],
-                        }
-                    );
-
-                    #[cfg(not(feature = "overlapped-lists"))]
-                    match data {
-                        Err(DeError::Custom(e)) => assert_eq!(e, "duplicate field `outer`"),
-                        e => panic!(
-                            r#"Expected Err(Custom("duplicate field `outer`")), got {:?}"#,
-                            e
-                        ),
-                    }
-                }
-            }
-
-            /// Deserialization of primitives slightly differs from deserialization
-            /// of complex types, so need to check this separately
-            #[test]
-            fn primitives() {
-                #[derive(Debug, PartialEq, Deserialize)]
-                struct List {
-                    item: Vec<usize>,
-                }
-
-                let data: List = from_str(
-                    r#"
-                    <root>
-                        <item>41</item>
-                        <item>42</item>
-                        <item>43</item>
-                    </root>
-                    "#,
-                )
-                .unwrap();
-
-                assert_eq!(
-                    data,
-                    List {
-                        item: vec![41, 42, 43],
-                    }
-                );
-
-                from_str::<List>(
-                    r#"
-                    <root>
-                        <item>41</item>
-                        <item><item>42</item></item>
-                        <item>43</item>
-                    </root>
-                    "#,
-                )
-                .unwrap_err();
-            }
-
-            /// This test ensures that composition of deserializer building blocks
-            /// plays well
-            #[test]
-            fn list_of_struct() {
-                #[derive(Debug, PartialEq, Default, Deserialize)]
-                #[serde(default)]
-                struct Struct {
-                    #[serde(rename = "@attribute")]
-                    attribute: Option<String>,
-                    element: Option<String>,
-                }
-
-                #[derive(Debug, PartialEq, Deserialize)]
-                struct List {
-                    item: Vec<Struct>,
-                }
-
-                let data: List = from_str(
-                    r#"
-                    <root>
-                        <item/>
-                        <item attribute="value"/>
-                        <item>
-                            <element>value</element>
-                        </item>
-                        <item attribute="value">
-                            <element>value</element>
-                        </item>
-                    </root>
-                    "#,
-                )
-                .unwrap();
-                assert_eq!(
-                    data,
-                    List {
-                        item: vec![
-                            Struct {
-                                attribute: None,
-                                element: None,
-                            },
-                            Struct {
-                                attribute: Some("value".to_string()),
-                                element: None,
-                            },
-                            Struct {
-                                attribute: None,
-                                element: Some("value".to_string()),
-                            },
-                            Struct {
-                                attribute: Some("value".to_string()),
-                                element: Some("value".to_string()),
-                            },
-                        ],
-                    }
-                );
-            }
-
-            /// Checks that sequences represented by elements can contain sequences,
-            /// represented by `xs:list`s
-            mod xs_list {
-                use super::*;
-                use pretty_assertions::assert_eq;
-
-                #[derive(Debug, Deserialize, PartialEq)]
-                struct List {
-                    /// Outer list mapped to elements, inner -- to `xs:list`.
-                    ///
-                    /// `#[serde(default)]` is required to correctly deserialize
-                    /// empty sequence, because without elements the field
-                    /// also is missing and derived `Deserialize` implementation
-                    /// would complain about that unless field is marked as
-                    /// `default`.
-                    #[serde(default)]
-                    item: Vec<Vec<String>>,
-                }
-
-                /// Special case: zero elements
-                #[test]
-                fn zero() {
-                    let data: List = from_str(
-                        r#"
-                        <root>
-                        </root>
-                        "#,
-                    )
-                    .unwrap();
-
-                    assert_eq!(data, List { item: vec![] });
-                }
-
-                /// Special case: one element
-                #[test]
-                fn one() {
-                    let data: List = from_str(
-                        r#"
-                        <root>
-                            <item>first list</item>
-                        </root>
-                        "#,
-                    )
-                    .unwrap();
-
-                    assert_eq!(
-                        data,
-                        List {
-                            item: vec![vec!["first".to_string(), "list".to_string()]]
-                        }
-                    );
-                }
-
-                /// Special case: outer list is always mapped to an elements sequence,
-                /// not to an `xs:list`
-                #[test]
-                fn element() {
-                    #[derive(Debug, Deserialize, PartialEq)]
-                    struct List {
-                        /// List mapped to elements, inner -- to `xs:list`.
-                        ///
-                        /// `#[serde(default)]` is not required, because correct
-                        /// XML will always contains at least 1 element.
-                        item: Vec<String>,
-                    }
-
-                    let data: List = from_str(
-                        r#"
-                        <root>
-                            <item>first item</item>
-                        </root>
-                        "#,
-                    )
-                    .unwrap();
-
-                    assert_eq!(
-                        data,
-                        List {
-                            item: vec!["first item".to_string()]
-                        }
-                    );
-                }
-
-                /// This tests demonstrates, that for `$value` field (`list`) actual
-                /// name of XML element (`item`) does not matter. That allows list
-                /// item to be an enum, where tag name determines enum variant
-                #[test]
-                fn many() {
-                    let data: List = from_str(
-                        r#"
-                        <root>
-                            <item>first list</item>
-                            <item>second list</item>
-                        </root>
-                        "#,
-                    )
-                    .unwrap();
-
-                    assert_eq!(
-                        data,
-                        List {
-                            item: vec![
-                                vec!["first".to_string(), "list".to_string()],
-                                vec!["second".to_string(), "list".to_string()],
-                            ]
-                        }
-                    );
-                }
-            }
-        }
-    }
-
-    /// Check that sequences inside element can be deserialized.
-    /// In terms of serde this is a sequence flatten into the struct:
-    ///
-    /// ```ignore
-    /// struct Root {
-    ///   #[serde(flatten)]
-    ///   items: Vec<T>,
-    /// }
-    /// ```
-    /// except that fact that this is not supported nowadays
-    /// (https://github.com/serde-rs/serde/issues/1905)
-    ///
-    /// Because this is very frequently used pattern in the XML, quick-xml
-    /// have a workaround for this. If a field will have a special name `$value`
-    /// then any `xs:element`s in the `xs:sequence` / `xs:all`, except that
-    /// which name matches the struct name, will be associated with this field:
-    ///
-    /// ```ignore
-    /// struct Root {
-    ///   field: U,
-    ///   #[serde(rename = "$value")]
-    ///   items: Vec<Enum>,
-    /// }
-    /// ```
-    /// In this example `<field>` tag will be associated with a `field` field,
-    /// but all other tags will be associated with an `items` field. Disadvantages
-    /// of this approach that you can have only one field, but usually you don't
-    /// want more
-    mod variable_name {
-        use super::*;
-        use serde::de::{Deserializer, EnumAccess, VariantAccess, Visitor};
-        use std::fmt::{self, Formatter};
-
-        // NOTE: Derive could be possible once https://github.com/serde-rs/serde/issues/2126 is resolved
-        macro_rules! impl_deserialize_choice {
-            ($name:ident : $(($field:ident, $field_name:literal)),*) => {
-                impl<'de> Deserialize<'de> for $name {
-                    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-                    where
-                        D: Deserializer<'de>,
-                    {
-                        #[derive(Deserialize)]
-                        #[serde(field_identifier)]
-                        #[serde(rename_all = "kebab-case")]
-                        enum Tag {
-                            $($field,)*
-                            Other(String),
-                        }
-
-                        struct EnumVisitor;
-                        impl<'de> Visitor<'de> for EnumVisitor {
-                            type Value = $name;
-
-                            fn expecting(&self, f: &mut Formatter) -> fmt::Result {
-                                f.write_str("enum ")?;
-                                f.write_str(stringify!($name))
-                            }
-
-                            fn visit_enum<A>(self, data: A) -> Result<Self::Value, A::Error>
-                            where
-                                A: EnumAccess<'de>,
-                            {
-                                match data.variant()? {
-                                    $(
-                                        (Tag::$field, variant) => variant.unit_variant().map(|_| $name::$field),
-                                    )*
-                                    (Tag::Other(t), v) => v.unit_variant().map(|_| $name::Other(t)),
-                                }
-                            }
-                        }
-
-                        const VARIANTS: &'static [&'static str] = &[
-                            $($field_name,)*
-                            "<any other tag>"
-                        ];
-                        deserializer.deserialize_enum(stringify!($name), VARIANTS, EnumVisitor)
-                    }
-                }
-            };
-        }
-
-        /// Type that can be deserialized from `<one>`, `<two>`, or any other element
-        #[derive(Debug, PartialEq)]
-        enum Choice {
-            One,
-            Two,
-            /// Any other tag name except `One` or `Two`, name of tag stored inside variant
-            Other(String),
-        }
-        impl_deserialize_choice!(Choice: (One, "one"), (Two, "two"));
-
-        /// Type that can be deserialized from `<first>`, `<second>`, or any other element
-        #[derive(Debug, PartialEq)]
-        enum Choice2 {
-            First,
-            Second,
-            /// Any other tag name except `First` or `Second`, name of tag stored inside variant
-            Other(String),
-        }
-        impl_deserialize_choice!(Choice2: (First, "first"), (Second, "second"));
-
-        /// Type that can be deserialized from `<one>`, `<two>`, or any other element.
-        /// Used for `primitives` tests
-        #[derive(Debug, PartialEq, Deserialize)]
-        #[serde(rename_all = "kebab-case")]
-        enum Choice3 {
-            One(usize),
-            Two(String),
-            #[serde(other)]
-            Other,
-        }
-
-        #[derive(Debug, PartialEq, Deserialize)]
-        #[serde(rename_all = "kebab-case")]
-        enum Choice4 {
-            One {
-                inner: [(); 1],
-            },
-            Two {
-                inner: [(); 1],
-            },
-            #[serde(other)]
-            Other,
-        }
-
-        /// This module contains tests where size of the list have a compile-time size
-        mod fixed_size {
-            use super::*;
-            use pretty_assertions::assert_eq;
-
-            #[derive(Debug, PartialEq, Deserialize)]
-            struct List {
-                #[serde(rename = "$value")]
-                item: [Choice; 3],
-            }
-
-            /// Simple case: count of elements matches expected size of sequence,
-            /// each element has the same name. Successful deserialization expected
-            #[test]
-            fn simple() {
-                let data: List = from_str(
-                    r#"
-                    <root>
-                        <one/>
-                        <two/>
-                        <three/>
-                    </root>
-                    "#,
-                )
-                .unwrap();
-
-                assert_eq!(
-                    data,
-                    List {
-                        item: [Choice::One, Choice::Two, Choice::Other("three".into())],
-                    }
-                );
-            }
-
-            /// Special case: empty sequence
-            #[test]
-            #[ignore = "it is impossible to distinguish between missed field and empty list: use `Option<>` or #[serde(default)]"]
-            fn empty() {
-                #[derive(Debug, PartialEq, Deserialize)]
-                struct List {
-                    #[serde(rename = "$value")]
-                    item: [Choice; 0],
-                }
-
-                from_str::<List>(r#"<root></root>"#).unwrap();
-                from_str::<List>(r#"<root/>"#).unwrap();
-            }
-
-            /// Special case: one-element sequence
-            #[test]
-            fn one_element() {
-                #[derive(Debug, PartialEq, Deserialize)]
-                struct List {
-                    #[serde(rename = "$value")]
-                    item: [Choice; 1],
-                }
-
-                let data: List = from_str(
-                    r#"
-                    <root>
-                        <one/>
-                    </root>
-                    "#,
-                )
-                .unwrap();
-
-                assert_eq!(
-                    data,
-                    List {
-                        item: [Choice::One],
-                    }
-                );
-            }
-
-            /// Fever elements than expected size of sequence, each element has
-            /// the same name. Failure expected
-            #[test]
-            fn fever_elements() {
-                from_str::<List>(
-                    r#"
-                    <root>
-                        <one/>
-                        <two/>
-                    </root>
-                    "#,
-                )
-                .unwrap_err();
-            }
-
-            /// More elements than expected size of sequence, each element has
-            /// the same name. Failure expected. If you wish to ignore excess
-            /// elements, use the special type, that consume as much elements
-            /// as possible, but ignores excess elements
-            #[test]
-            fn excess_elements() {
-                from_str::<List>(
-                    r#"
-                    <root>
-                        <one/>
-                        <two/>
-                        <three/>
-                        <four/>
-                    </root>
-                    "#,
-                )
-                .unwrap_err();
-            }
-
-            #[test]
-            fn mixed_content() {
-                #[derive(Debug, PartialEq, Deserialize)]
-                struct List {
-                    /// Text and CDATA represents a one logical text item
-                    #[serde(rename = "$value")]
-                    item: [(); 2],
-                }
-
-                from_str::<List>(
-                    r#"
-                    <root>
-                        <element/>
-                        text
-                        <![CDATA[cdata]]>
-                    </root>
-                    "#,
-                )
-                .unwrap();
-            }
-
-            // There cannot be unknown items, because any tag name is accepted
-
-            /// In those tests non-sequential field is defined in the struct
-            /// before sequential, so it will be deserialized before the list.
-            /// That struct should be deserialized from the XML where these
-            /// fields comes in an arbitrary order
-            mod field_before_list {
-                use super::*;
-                use pretty_assertions::assert_eq;
-
-                #[derive(Debug, PartialEq, Deserialize)]
-                struct Root {
-                    node: (),
-                    #[serde(rename = "$value")]
-                    item: [Choice; 3],
-                }
-
-                #[test]
-                fn before() {
-                    let data: Root = from_str(
-                        r#"
-                        <root>
-                            <node/>
-                            <one/>
-                            <two/>
-                            <three/>
-                        </root>
-                        "#,
-                    )
-                    .unwrap();
-
-                    assert_eq!(
-                        data,
-                        Root {
-                            node: (),
-                            item: [Choice::One, Choice::Two, Choice::Other("three".into())],
-                        }
-                    );
-                }
-
-                #[test]
-                fn after() {
-                    let data: Root = from_str(
-                        r#"
-                        <root>
-                            <one/>
-                            <two/>
-                            <three/>
-                            <node/>
-                        </root>
-                        "#,
-                    )
-                    .unwrap();
-
-                    assert_eq!(
-                        data,
-                        Root {
-                            node: (),
-                            item: [Choice::One, Choice::Two, Choice::Other("three".into())],
-                        }
-                    );
-                }
-
-                #[test]
-                fn overlapped() {
-                    let data = from_str::<Root>(
-                        r#"
-                        <root>
-                            <one/>
-                            <node/>
-                            <two/>
-                            <three/>
-                        </root>
-                        "#,
-                    );
-
-                    #[cfg(feature = "overlapped-lists")]
-                    assert_eq!(
-                        data.unwrap(),
-                        Root {
-                            node: (),
-                            item: [Choice::One, Choice::Two, Choice::Other("three".into())],
-                        }
-                    );
-
-                    #[cfg(not(feature = "overlapped-lists"))]
-                    match data {
-                        Err(DeError::Custom(e)) => {
-                            assert_eq!(e, "invalid length 1, expected an array of length 3")
-                        }
-                        e => panic!(
-                            r#"Expected Err(Custom("invalid length 1, expected an array of length 3")), got {:?}"#,
-                            e
-                        ),
-                    }
-                }
-
-                /// Test for https://github.com/tafia/quick-xml/issues/435
-                #[test]
-                fn overlapped_with_nested_list() {
-                    #[derive(Debug, PartialEq, Deserialize)]
-                    struct Root {
-                        node: (),
-                        #[serde(rename = "$value")]
-                        item: [Choice4; 3],
-                    }
-
-                    let data = from_str::<Root>(
-                        r#"
-                        <root>
-                            <one><inner/></one>
-                            <node/>
-                            <two><inner/></two>
-                            <three><inner/></three>
-                        </root>
-                        "#,
-                    );
-
-                    #[cfg(feature = "overlapped-lists")]
-                    assert_eq!(
-                        data.unwrap(),
-                        Root {
-                            node: (),
-                            item: [
-                                Choice4::One { inner: [()] },
-                                Choice4::Two { inner: [()] },
-                                Choice4::Other,
-                            ],
-                        }
-                    );
-
-                    #[cfg(not(feature = "overlapped-lists"))]
-                    match data {
-                        Err(DeError::Custom(e)) => {
-                            assert_eq!(e, "invalid length 1, expected an array of length 3")
-                        }
-                        e => panic!(
-                            r#"Expected Err(Custom("invalid length 1, expected an array of length 3")), got {:?}"#,
-                            e
-                        ),
-                    }
-                }
-            }
-
-            /// In those tests non-sequential field is defined in the struct
-            /// after sequential, so it will be deserialized after the list.
-            /// That struct should be deserialized from the XML where these
-            /// fields comes in an arbitrary order
-            mod field_after_list {
-                use super::*;
-                use pretty_assertions::assert_eq;
-
-                #[derive(Debug, PartialEq, Deserialize)]
-                struct Root {
-                    #[serde(rename = "$value")]
-                    item: [Choice; 3],
-                    node: (),
-                }
-
-                #[test]
-                fn before() {
-                    let data: Root = from_str(
-                        r#"
-                        <root>
-                            <node/>
-                            <one/>
-                            <two/>
-                            <three/>
-                        </root>
-                        "#,
-                    )
-                    .unwrap();
-
-                    assert_eq!(
-                        data,
-                        Root {
-                            item: [Choice::One, Choice::Two, Choice::Other("three".into())],
-                            node: (),
-                        }
-                    );
-                }
-
-                #[test]
-                fn after() {
-                    let data: Root = from_str(
-                        r#"
-                        <root>
-                            <one/>
-                            <two/>
-                            <three/>
-                            <node/>
-                        </root>
-                        "#,
-                    )
-                    .unwrap();
-
-                    assert_eq!(
-                        data,
-                        Root {
-                            item: [Choice::One, Choice::Two, Choice::Other("three".into())],
-                            node: (),
-                        }
-                    );
-                }
-
-                #[test]
-                fn overlapped() {
-                    let data = from_str::<Root>(
-                        r#"
-                        <root>
-                            <one/>
-                            <node/>
-                            <two/>
-                            <three/>
-                        </root>
-                        "#,
-                    );
-
-                    #[cfg(feature = "overlapped-lists")]
-                    assert_eq!(
-                        data.unwrap(),
-                        Root {
-                            item: [Choice::One, Choice::Two, Choice::Other("three".into())],
-                            node: (),
-                        }
-                    );
-
-                    #[cfg(not(feature = "overlapped-lists"))]
-                    match data {
-                        Err(DeError::Custom(e)) => {
-                            assert_eq!(e, "invalid length 1, expected an array of length 3")
-                        }
-                        e => panic!(
-                            r#"Expected Err(Custom("invalid length 1, expected an array of length 3")), got {:?}"#,
-                            e
-                        ),
-                    }
-                }
-
-                /// Test for https://github.com/tafia/quick-xml/issues/435
-                #[test]
-                fn overlapped_with_nested_list() {
-                    #[derive(Debug, PartialEq, Deserialize)]
-                    struct Root {
-                        #[serde(rename = "$value")]
-                        item: [Choice4; 3],
-                        node: (),
-                    }
-
-                    let data = from_str::<Root>(
-                        r#"
-                        <root>
-                            <one><inner/></one>
-                            <node/>
-                            <two><inner/></two>
-                            <three><inner/></three>
-                        </root>
-                        "#,
-                    );
-
-                    #[cfg(feature = "overlapped-lists")]
-                    assert_eq!(
-                        data.unwrap(),
-                        Root {
-                            item: [
-                                Choice4::One { inner: [()] },
-                                Choice4::Two { inner: [()] },
-                                Choice4::Other,
-                            ],
-                            node: (),
-                        }
-                    );
-
-                    #[cfg(not(feature = "overlapped-lists"))]
-                    match data {
-                        Err(DeError::Custom(e)) => {
-                            assert_eq!(e, "invalid length 1, expected an array of length 3")
-                        }
-                        e => panic!(
-                            r#"Expected Err(Custom("invalid length 1, expected an array of length 3")), got {:?}"#,
-                            e
-                        ),
-                    }
-                }
-            }
-
-            /// In those tests two lists are deserialized simultaneously.
-            /// Lists should be deserialized even when them overlaps
-            mod two_lists {
-                use super::*;
-
-                /// A field with a variable-name items defined before a field with a fixed-name
-                /// items
-                mod choice_and_fixed {
-                    use super::*;
-                    use pretty_assertions::assert_eq;
-
-                    #[derive(Debug, PartialEq, Deserialize)]
-                    struct Pair {
-                        #[serde(rename = "$value")]
-                        item: [Choice; 3],
-                        element: [(); 2],
-                    }
-
-                    /// A list with fixed-name elements located before a list with variable-name
-                    /// elements in an XML
-                    #[test]
-                    fn fixed_before() {
-                        let data: Pair = from_str(
-                            r#"
-                            <root>
-                                <element/>
-                                <element/>
-                                <one/>
-                                <two/>
-                                <three/>
-                            </root>
-                            "#,
-                        )
-                        .unwrap();
-
-                        assert_eq!(
-                            data,
-                            Pair {
-                                item: [Choice::One, Choice::Two, Choice::Other("three".into())],
-                                element: [(), ()],
-                            }
-                        );
-                    }
-
-                    /// A list with fixed-name elements located after a list with variable-name
-                    /// elements in an XML
-                    #[test]
-                    fn fixed_after() {
-                        let data: Pair = from_str(
-                            r#"
-                            <root>
-                                <one/>
-                                <two/>
-                                <three/>
-                                <element/>
-                                <element/>
-                            </root>
-                            "#,
-                        )
-                        .unwrap();
-
-                        assert_eq!(
-                            data,
-                            Pair {
-                                item: [Choice::One, Choice::Two, Choice::Other("three".into())],
-                                element: [(), ()],
-                            }
-                        );
-                    }
-
-                    mod overlapped {
-                        use super::*;
-                        use pretty_assertions::assert_eq;
-
-                        #[derive(Debug, PartialEq, Deserialize)]
-                        struct Root {
-                            #[serde(rename = "$value")]
-                            item: [Choice4; 3],
-                            element: [(); 2],
-                        }
-
-                        /// A list with fixed-name elements are mixed with a list with variable-name
-                        /// elements in an XML, and the first element is a fixed-name one
-                        #[test]
-                        fn fixed_before() {
-                            let data = from_str::<Pair>(
-                                r#"
-                                <root>
-                                    <element/>
-                                    <one/>
-                                    <two/>
-                                    <element/>
-                                    <three/>
-                                </root>
-                                "#,
-                            );
-
-                            #[cfg(feature = "overlapped-lists")]
-                            assert_eq!(
-                                data.unwrap(),
-                                Pair {
-                                    item: [Choice::One, Choice::Two, Choice::Other("three".into())],
-                                    element: [(), ()],
-                                }
-                            );
-
-                            #[cfg(not(feature = "overlapped-lists"))]
-                            match data {
-                                Err(DeError::Custom(e)) => {
-                                    assert_eq!(e, "invalid length 1, expected an array of length 2")
-                                }
-                                e => panic!(
-                                    r#"Expected Err(Custom("invalid length 1, expected an array of length 2")), got {:?}"#,
-                                    e
-                                ),
-                            }
-                        }
-
-                        /// A list with fixed-name elements are mixed with a list with variable-name
-                        /// elements in an XML, and the first element is a variable-name one
-                        #[test]
-                        fn fixed_after() {
-                            let data = from_str::<Pair>(
-                                r#"
-                                <root>
-                                    <one/>
-                                    <element/>
-                                    <two/>
-                                    <three/>
-                                    <element/>
-                                </root>
-                                "#,
-                            );
-
-                            #[cfg(feature = "overlapped-lists")]
-                            assert_eq!(
-                                data.unwrap(),
-                                Pair {
-                                    item: [Choice::One, Choice::Two, Choice::Other("three".into())],
-                                    element: [(), ()],
-                                }
-                            );
-
-                            #[cfg(not(feature = "overlapped-lists"))]
-                            match data {
-                                Err(DeError::Custom(e)) => {
-                                    assert_eq!(e, "invalid length 1, expected an array of length 3")
-                                }
-                                e => panic!(
-                                    r#"Expected Err(Custom("invalid length 1, expected an array of length 3")), got {:?}"#,
-                                    e
-                                ),
-                            }
-                        }
-
-                        /// Test for https://github.com/tafia/quick-xml/issues/435
-                        #[test]
-                        fn with_nested_list_fixed_before() {
-                            let data = from_str::<Root>(
-                                r#"
-                                <root>
-                                    <element/>
-                                    <one><inner/></one>
-                                    <two><inner/></two>
-                                    <element/>
-                                    <three><inner/></three>
-                                </root>
-                                "#,
-                            );
-
-                            #[cfg(feature = "overlapped-lists")]
-                            assert_eq!(
-                                data.unwrap(),
-                                Root {
-                                    item: [
-                                        Choice4::One { inner: [()] },
-                                        Choice4::Two { inner: [()] },
-                                        Choice4::Other,
-                                    ],
-                                    element: [(); 2],
-                                }
-                            );
-
-                            #[cfg(not(feature = "overlapped-lists"))]
-                            match data {
-                                Err(DeError::Custom(e)) => {
-                                    assert_eq!(e, "invalid length 1, expected an array of length 2")
-                                }
-                                e => panic!(
-                                    r#"Expected Err(Custom("invalid length 1, expected an array of length 2")), got {:?}"#,
-                                    e
-                                ),
-                            }
-                        }
-
-                        /// Test for https://github.com/tafia/quick-xml/issues/435
-                        #[test]
-                        fn with_nested_list_fixed_after() {
-                            let data = from_str::<Root>(
-                                r#"
-                                <root>
-                                    <one><inner/></one>
-                                    <element/>
-                                    <two><inner/></two>
-                                    <three><inner/></three>
-                                    <element/>
-                                </root>
-                                "#,
-                            );
-
-                            #[cfg(feature = "overlapped-lists")]
-                            assert_eq!(
-                                data.unwrap(),
-                                Root {
-                                    item: [
-                                        Choice4::One { inner: [()] },
-                                        Choice4::Two { inner: [()] },
-                                        Choice4::Other,
-                                    ],
-                                    element: [(); 2],
-                                }
-                            );
-
-                            #[cfg(not(feature = "overlapped-lists"))]
-                            match data {
-                                Err(DeError::Custom(e)) => {
-                                    assert_eq!(e, "invalid length 1, expected an array of length 3")
-                                }
-                                e => panic!(
-                                    r#"Expected Err(Custom("invalid length 1, expected an array of length 3")), got {:?}"#,
-                                    e
-                                ),
-                            }
-                        }
-                    }
-                }
-
-                /// A field with a variable-name items defined after a field with a fixed-name
-                /// items
-                mod fixed_and_choice {
-                    use super::*;
-                    use pretty_assertions::assert_eq;
-
-                    #[derive(Debug, PartialEq, Deserialize)]
-                    struct Pair {
-                        element: [(); 2],
-                        #[serde(rename = "$value")]
-                        item: [Choice; 3],
-                    }
-
-                    /// A list with fixed-name elements located before a list with variable-name
-                    /// elements in an XML
-                    #[test]
-                    fn fixed_before() {
-                        let data: Pair = from_str(
-                            r#"
-                            <root>
-                                <element/>
-                                <element/>
-                                <one/>
-                                <two/>
-                                <three/>
-                            </root>
-                            "#,
-                        )
-                        .unwrap();
-
-                        assert_eq!(
-                            data,
-                            Pair {
-                                item: [Choice::One, Choice::Two, Choice::Other("three".into())],
-                                element: [(), ()],
-                            }
-                        );
-                    }
-
-                    /// A list with fixed-name elements located after a list with variable-name
-                    /// elements in an XML
-                    #[test]
-                    fn fixed_after() {
-                        let data: Pair = from_str(
-                            r#"
-                            <root>
-                                <one/>
-                                <two/>
-                                <three/>
-                                <element/>
-                                <element/>
-                            </root>
-                            "#,
-                        )
-                        .unwrap();
-
-                        assert_eq!(
-                            data,
-                            Pair {
-                                item: [Choice::One, Choice::Two, Choice::Other("three".into())],
-                                element: [(), ()],
-                            }
-                        );
-                    }
-
-                    mod overlapped {
-                        use super::*;
-                        use pretty_assertions::assert_eq;
-
-                        #[derive(Debug, PartialEq, Deserialize)]
-                        struct Root {
-                            element: [(); 2],
-                            #[serde(rename = "$value")]
-                            item: [Choice4; 3],
-                        }
-
-                        /// A list with fixed-name elements are mixed with a list with variable-name
-                        /// elements in an XML, and the first element is a fixed-name one
-                        #[test]
-                        fn fixed_before() {
-                            let data = from_str::<Pair>(
-                                r#"
-                                <root>
-                                    <element/>
-                                    <one/>
-                                    <two/>
-                                    <element/>
-                                    <three/>
-                                </root>
-                                "#,
-                            );
-
-                            #[cfg(feature = "overlapped-lists")]
-                            assert_eq!(
-                                data.unwrap(),
-                                Pair {
-                                    item: [Choice::One, Choice::Two, Choice::Other("three".into())],
-                                    element: [(), ()],
-                                }
-                            );
-
-                            #[cfg(not(feature = "overlapped-lists"))]
-                            match data {
-                                Err(DeError::Custom(e)) => {
-                                    assert_eq!(e, "invalid length 1, expected an array of length 2")
-                                }
-                                e => panic!(
-                                    r#"Expected Err(Custom("invalid length 1, expected an array of length 2")), got {:?}"#,
-                                    e
-                                ),
-                            }
-                        }
-
-                        /// A list with fixed-name elements are mixed with a list with variable-name
-                        /// elements in an XML, and the first element is a variable-name one
-                        #[test]
-                        fn fixed_after() {
-                            let data = from_str::<Pair>(
-                                r#"
-                                <root>
-                                    <one/>
-                                    <element/>
-                                    <two/>
-                                    <three/>
-                                    <element/>
-                                </root>
-                                "#,
-                            );
-
-                            #[cfg(feature = "overlapped-lists")]
-                            assert_eq!(
-                                data.unwrap(),
-                                Pair {
-                                    item: [Choice::One, Choice::Two, Choice::Other("three".into())],
-                                    element: [(), ()],
-                                }
-                            );
-
-                            #[cfg(not(feature = "overlapped-lists"))]
-                            match data {
-                                Err(DeError::Custom(e)) => {
-                                    assert_eq!(e, "invalid length 1, expected an array of length 3")
-                                }
-                                e => panic!(
-                                    r#"Expected Err(Custom("invalid length 1, expected an array of length 3")), got {:?}"#,
-                                    e
-                                ),
-                            }
-                        }
-
-                        /// Test for https://github.com/tafia/quick-xml/issues/435
-                        #[test]
-                        fn with_nested_list_fixed_before() {
-                            let data = from_str::<Root>(
-                                r#"
-                                <root>
-                                    <element/>
-                                    <one><inner/></one>
-                                    <two><inner/></two>
-                                    <element/>
-                                    <three><inner/></three>
-                                </root>
-                                "#,
-                            );
-
-                            #[cfg(feature = "overlapped-lists")]
-                            assert_eq!(
-                                data.unwrap(),
-                                Root {
-                                    element: [(); 2],
-                                    item: [
-                                        Choice4::One { inner: [()] },
-                                        Choice4::Two { inner: [()] },
-                                        Choice4::Other,
-                                    ],
-                                }
-                            );
-
-                            #[cfg(not(feature = "overlapped-lists"))]
-                            match data {
-                                Err(DeError::Custom(e)) => {
-                                    assert_eq!(e, "invalid length 1, expected an array of length 2")
-                                }
-                                e => panic!(
-                                    r#"Expected Err(Custom("invalid length 1, expected an array of length 2")), got {:?}"#,
-                                    e
-                                ),
-                            }
-                        }
-
-                        /// Test for https://github.com/tafia/quick-xml/issues/435
-                        #[test]
-                        fn with_nested_list_fixed_after() {
-                            let data = from_str::<Root>(
-                                r#"
-                                <root>
-                                    <one><inner/></one>
-                                    <element/>
-                                    <two><inner/></two>
-                                    <three><inner/></three>
-                                    <element/>
-                                </root>
-                                "#,
-                            );
-
-                            #[cfg(feature = "overlapped-lists")]
-                            assert_eq!(
-                                data.unwrap(),
-                                Root {
-                                    element: [(); 2],
-                                    item: [
-                                        Choice4::One { inner: [()] },
-                                        Choice4::Two { inner: [()] },
-                                        Choice4::Other,
-                                    ],
-                                }
-                            );
-
-                            #[cfg(not(feature = "overlapped-lists"))]
-                            match data {
-                                Err(DeError::Custom(e)) => {
-                                    assert_eq!(e, "invalid length 1, expected an array of length 3")
-                                }
-                                e => panic!(
-                                    r#"Expected Err(Custom("invalid length 1, expected an array of length 3")), got {:?}"#,
-                                    e
-                                ),
-                            }
-                        }
-                    }
-                }
-
-                /// Tests are ignored, but exists to show a problem.
-                /// May be it will be solved in the future
-                mod choice_and_choice {
-                    use super::*;
-                    use pretty_assertions::assert_eq;
-
-                    #[derive(Debug, PartialEq, Deserialize)]
-                    struct Pair {
-                        #[serde(rename = "$value")]
-                        item: [Choice; 3],
-                        // Actually, we cannot rename both fields to `$value`, which is now
-                        // required to indicate, that field accepts elements with any name
-                        #[serde(rename = "$value")]
-                        element: [Choice2; 2],
-                    }
-
-                    #[test]
-                    #[ignore = "There is no way to associate XML elements with `item` or `element` without extra knowledge from type"]
-                    fn splitted() {
-                        let data: Pair = from_str(
-                            r#"
-                            <root>
-                                <first/>
-                                <second/>
-                                <one/>
-                                <two/>
-                                <three/>
-                            </root>
-                            "#,
-                        )
-                        .unwrap();
-
-                        assert_eq!(
-                            data,
-                            Pair {
-                                item: [Choice::One, Choice::Two, Choice::Other("three".into())],
-                                element: [Choice2::First, Choice2::Second],
-                            }
-                        );
-                    }
-
-                    #[test]
-                    #[ignore = "There is no way to associate XML elements with `item` or `element` without extra knowledge from type"]
-                    fn overlapped() {
-                        let data = from_str::<Pair>(
-                            r#"
-                            <root>
-                                <one/>
-                                <first/>
-                                <two/>
-                                <second/>
-                                <three/>
-                            </root>
-                            "#,
-                        );
-
-                        #[cfg(feature = "overlapped-lists")]
-                        assert_eq!(
-                            data.unwrap(),
-                            Pair {
-                                item: [Choice::One, Choice::Two, Choice::Other("three".into())],
-                                element: [Choice2::First, Choice2::Second],
-                            }
-                        );
-
-                        #[cfg(not(feature = "overlapped-lists"))]
-                        match data {
-                            Err(DeError::Custom(e)) => {
-                                assert_eq!(e, "invalid length 1, expected an array of length 3")
-                            }
-                            e => panic!(
-                                r#"Expected Err(Custom("invalid length 1, expected an array of length 3")), got {:?}"#,
-                                e
-                            ),
-                        }
-                    }
-                }
-            }
-
-            /// Deserialization of primitives slightly differs from deserialization
-            /// of complex types, so need to check this separately
-            #[test]
-            fn primitives() {
-                #[derive(Debug, PartialEq, Deserialize)]
-                struct List {
-                    #[serde(rename = "$value")]
-                    item: [Choice3; 3],
-                }
-
-                let data: List = from_str(
-                    r#"
-                    <root>
-                        <one>41</one>
-                        <two>42</two>
-                        <three>43</three>
-                    </root>
-                    "#,
-                )
-                .unwrap();
-
-                assert_eq!(
-                    data,
-                    List {
-                        item: [
-                            Choice3::One(41),
-                            Choice3::Two("42".to_string()),
-                            Choice3::Other,
-                        ],
-                    }
-                );
-
-                from_str::<List>(
-                    r#"
-                    <root>
-                        <one>41</one>
-                        <two><item>42</item></two>
-                        <three>43</three>
-                    </root>
-                    "#,
-                )
-                .unwrap_err();
-            }
-
-            /// Checks that sequences represented by elements can contain sequences,
-            /// represented by `xs:list`s
-            mod xs_list {
-                use super::*;
-                use pretty_assertions::assert_eq;
-
-                /// Special case: zero elements
-                #[test]
-                fn zero() {
-                    #[derive(Debug, Deserialize, PartialEq)]
-                    struct List {
-                        /// Outer list mapped to elements, inner -- to `xs:list`.
-                        ///
-                        /// `#[serde(default)]` is required to correctly deserialize
-                        /// empty sequence, because without elements the field
-                        /// also is missing and derived `Deserialize` implementation
-                        /// would complain about that unless field is marked as
-                        /// `default`.
-                        #[serde(default)]
-                        #[serde(rename = "$value")]
-                        element: [Vec<String>; 0],
-                    }
-
-                    let data: List = from_str(
-                        r#"
-                        <root>
-                        </root>
-                        "#,
-                    )
-                    .unwrap();
-
-                    assert_eq!(data, List { element: [] });
-                }
-
-                /// Special case: one element
-                #[test]
-                fn one() {
-                    #[derive(Debug, Deserialize, PartialEq)]
-                    struct List {
-                        /// Outer list mapped to elements, inner -- to `xs:list`.
-                        ///
-                        /// `#[serde(default)]` is not required, because correct
-                        /// XML will always contains at least 1 element.
-                        #[serde(rename = "$value")]
-                        element: [Vec<String>; 1],
-                    }
-
-                    let data: List = from_str(
-                        r#"
-                        <root>
-                            <item>first list</item>
-                        </root>
-                        "#,
-                    )
-                    .unwrap();
-
-                    assert_eq!(
-                        data,
-                        List {
-                            element: [vec!["first".to_string(), "list".to_string()]]
-                        }
-                    );
-                }
-
-                /// Special case: outer list is always mapped to an elements sequence,
-                /// not to an `xs:list`
-                #[test]
-                fn element() {
-                    #[derive(Debug, Deserialize, PartialEq)]
-                    struct List {
-                        /// List mapped to elements, String -- to `xs:list`.
-                        ///
-                        /// `#[serde(default)]` is not required, because correct
-                        /// XML will always contains at least 1 element.
-                        #[serde(rename = "$value")]
-                        element: [String; 1],
-                    }
-
-                    let data: List = from_str(
-                        r#"
-                        <root>
-                            <item>first item</item>
-                        </root>
-                        "#,
-                    )
-                    .unwrap();
-
-                    assert_eq!(
-                        data,
-                        List {
-                            element: ["first item".to_string()]
-                        }
-                    );
-                }
-
-                #[test]
-                fn many() {
-                    #[derive(Debug, Deserialize, PartialEq)]
-                    struct List {
-                        /// Outer list mapped to elements, inner -- to `xs:list`
-                        #[serde(rename = "$value")]
-                        element: [Vec<String>; 2],
-                    }
-
-                    let data: List = from_str(
-                        r#"
-                        <root>
-                            <item>first list</item>
-                            <item>second list</item>
-                        </root>
-                        "#,
-                    )
-                    .unwrap();
-
-                    assert_eq!(
-                        data,
-                        List {
-                            element: [
-                                vec!["first".to_string(), "list".to_string()],
-                                vec!["second".to_string(), "list".to_string()],
-                            ]
-                        }
-                    );
-                }
-            }
-        }
-
-        /// This module contains tests where size of the list have an unspecified size
-        mod variable_size {
-            use super::*;
-            use pretty_assertions::assert_eq;
-
-            #[derive(Debug, PartialEq, Deserialize)]
-            struct List {
-                #[serde(rename = "$value")]
-                item: Vec<Choice>,
-            }
-
-            /// Simple case: count of elements matches expected size of sequence,
-            /// each element has the same name. Successful deserialization expected
-            #[test]
-            fn simple() {
-                let data: List = from_str(
-                    r#"
-                    <root>
-                        <one/>
-                        <two/>
-                        <three/>
-                    </root>
-                    "#,
-                )
-                .unwrap();
-
-                assert_eq!(
-                    data,
-                    List {
-                        item: vec![Choice::One, Choice::Two, Choice::Other("three".into())],
-                    }
-                );
-            }
-
-            /// Special case: empty sequence
-            #[test]
-            #[ignore = "it is impossible to distinguish between missed field and empty list: use `Option<>` or #[serde(default)]"]
-            fn empty() {
-                let data = from_str::<List>(r#"<root></root>"#).unwrap();
-                assert_eq!(data, List { item: vec![] });
-
-                let data = from_str::<List>(r#"<root/>"#).unwrap();
-                assert_eq!(data, List { item: vec![] });
-            }
-
-            /// Special case: one-element sequence
-            #[test]
-            fn one_element() {
-                let data: List = from_str(
-                    r#"
-                    <root>
-                        <one/>
-                    </root>
-                    "#,
-                )
-                .unwrap();
-
-                assert_eq!(
-                    data,
-                    List {
-                        item: vec![Choice::One],
-                    }
-                );
-            }
-
-            #[test]
-            fn mixed_content() {
-                #[derive(Debug, PartialEq, Deserialize)]
-                struct List {
-                    #[serde(rename = "$value")]
-                    item: Vec<()>,
-                }
-
-                let data: List = from_str(
-                    r#"
-                    <root>
-                        <element/>
-                        text
-                        <![CDATA[cdata]]>
-                    </root>
-                    "#,
-                )
-                .unwrap();
-
-                assert_eq!(
-                    data,
-                    List {
-                        // Text and CDATA represents a one logical text item
-                        item: vec![(), ()],
-                    }
-                );
-            }
-
-            // There cannot be unknown items, because any tag name is accepted
-
-            /// In those tests non-sequential field is defined in the struct
-            /// before sequential, so it will be deserialized before the list.
-            /// That struct should be deserialized from the XML where these
-            /// fields comes in an arbitrary order
-            mod field_before_list {
-                use super::*;
-                use pretty_assertions::assert_eq;
-
-                #[derive(Debug, PartialEq, Deserialize)]
-                struct Root {
-                    node: (),
-                    #[serde(rename = "$value")]
-                    item: Vec<Choice>,
-                }
-
-                #[test]
-                fn before() {
-                    let data: Root = from_str(
-                        r#"
-                        <root>
-                            <node/>
-                            <one/>
-                            <two/>
-                            <three/>
-                        </root>
-                        "#,
-                    )
-                    .unwrap();
-
-                    assert_eq!(
-                        data,
-                        Root {
-                            node: (),
-                            item: vec![Choice::One, Choice::Two, Choice::Other("three".into())],
-                        }
-                    );
-                }
-
-                #[test]
-                fn after() {
-                    let data: Root = from_str(
-                        r#"
-                        <root>
-                            <one/>
-                            <two/>
-                            <three/>
-                            <node/>
-                        </root>
-                        "#,
-                    )
-                    .unwrap();
-
-                    assert_eq!(
-                        data,
-                        Root {
-                            node: (),
-                            item: vec![Choice::One, Choice::Two, Choice::Other("three".into())],
-                        }
-                    );
-                }
-
-                #[test]
-                fn overlapped() {
-                    let data = from_str::<Root>(
-                        r#"
-                        <root>
-                            <one/>
-                            <node/>
-                            <two/>
-                            <three/>
-                        </root>
-                        "#,
-                    );
-
-                    #[cfg(feature = "overlapped-lists")]
-                    assert_eq!(
-                        data.unwrap(),
-                        Root {
-                            node: (),
-                            item: vec![Choice::One, Choice::Two, Choice::Other("three".into())],
-                        }
-                    );
-
-                    #[cfg(not(feature = "overlapped-lists"))]
-                    match data {
-                        Err(DeError::Custom(e)) => assert_eq!(e, "duplicate field `$value`"),
-                        e => panic!(
-                            r#"Expected Err(Custom("duplicate field `$value`")), got {:?}"#,
-                            e
-                        ),
-                    }
-                }
-
-                /// Test for https://github.com/tafia/quick-xml/issues/435
-                #[test]
-                fn overlapped_with_nested_list() {
-                    #[derive(Debug, PartialEq, Deserialize)]
-                    struct Root {
-                        node: (),
-                        #[serde(rename = "$value")]
-                        item: Vec<Choice4>,
-                    }
-
-                    let data = from_str::<Root>(
-                        r#"
-                        <root>
-                            <one><inner/></one>
-                            <node/>
-                            <two><inner/></two>
-                            <three><inner/></three>
-                        </root>
-                        "#,
-                    );
-
-                    #[cfg(feature = "overlapped-lists")]
-                    assert_eq!(
-                        data.unwrap(),
-                        Root {
-                            node: (),
-                            item: vec![
-                                Choice4::One { inner: [()] },
-                                Choice4::Two { inner: [()] },
-                                Choice4::Other,
-                            ],
-                        }
-                    );
-
-                    #[cfg(not(feature = "overlapped-lists"))]
-                    match data {
-                        Err(DeError::Custom(e)) => {
-                            assert_eq!(e, "duplicate field `$value`")
-                        }
-                        e => panic!(
-                            r#"Expected Err(Custom("duplicate field `$value`")), got {:?}"#,
-                            e
-                        ),
-                    }
-                }
-            }
-
-            /// In those tests non-sequential field is defined in the struct
-            /// after sequential, so it will be deserialized after the list.
-            /// That struct should be deserialized from the XML where these
-            /// fields comes in an arbitrary order
-            mod field_after_list {
-                use super::*;
-                use pretty_assertions::assert_eq;
-
-                #[derive(Debug, PartialEq, Deserialize)]
-                struct Root {
-                    #[serde(rename = "$value")]
-                    item: Vec<Choice>,
-                    node: (),
-                }
-
-                #[test]
-                fn before() {
-                    let data: Root = from_str(
-                        r#"
-                        <root>
-                            <node/>
-                            <one/>
-                            <two/>
-                            <three/>
-                        </root>
-                        "#,
-                    )
-                    .unwrap();
-
-                    assert_eq!(
-                        data,
-                        Root {
-                            item: vec![Choice::One, Choice::Two, Choice::Other("three".into())],
-                            node: (),
-                        }
-                    );
-                }
-
-                #[test]
-                fn after() {
-                    let data: Root = from_str(
-                        r#"
-                        <root>
-                            <one/>
-                            <two/>
-                            <three/>
-                            <node/>
-                        </root>
-                        "#,
-                    )
-                    .unwrap();
-
-                    assert_eq!(
-                        data,
-                        Root {
-                            item: vec![Choice::One, Choice::Two, Choice::Other("three".into())],
-                            node: (),
-                        }
-                    );
-                }
-
-                #[test]
-                fn overlapped() {
-                    let data = from_str::<Root>(
-                        r#"
-                        <root>
-                            <one/>
-                            <node/>
-                            <two/>
-                            <three/>
-                        </root>
-                        "#,
-                    );
-
-                    #[cfg(feature = "overlapped-lists")]
-                    assert_eq!(
-                        data.unwrap(),
-                        Root {
-                            item: vec![Choice::One, Choice::Two, Choice::Other("three".into())],
-                            node: (),
-                        }
-                    );
-
-                    #[cfg(not(feature = "overlapped-lists"))]
-                    match data {
-                        Err(DeError::Custom(e)) => assert_eq!(e, "duplicate field `$value`"),
-                        e => panic!(
-                            r#"Expected Err(Custom("duplicate field `$value`")), got {:?}"#,
-                            e
-                        ),
-                    }
-                }
-
-                /// Test for https://github.com/tafia/quick-xml/issues/435
-                #[test]
-                fn overlapped_with_nested_list() {
-                    #[derive(Debug, PartialEq, Deserialize)]
-                    struct Root {
-                        #[serde(rename = "$value")]
-                        item: Vec<Choice4>,
-                        node: (),
-                    }
-
-                    let data = from_str::<Root>(
-                        r#"
-                        <root>
-                            <one><inner/></one>
-                            <node/>
-                            <two><inner/></two>
-                            <three><inner/></three>
-                        </root>
-                        "#,
-                    );
-
-                    #[cfg(feature = "overlapped-lists")]
-                    assert_eq!(
-                        data.unwrap(),
-                        Root {
-                            item: vec![
-                                Choice4::One { inner: [()] },
-                                Choice4::Two { inner: [()] },
-                                Choice4::Other,
-                            ],
-                            node: (),
-                        }
-                    );
-
-                    #[cfg(not(feature = "overlapped-lists"))]
-                    match data {
-                        Err(DeError::Custom(e)) => {
-                            assert_eq!(e, "duplicate field `$value`")
-                        }
-                        e => panic!(
-                            r#"Expected Err(Custom("duplicate field `$value`")), got {:?}"#,
-                            e
-                        ),
-                    }
-                }
-            }
-
-            /// In those tests two lists are deserialized simultaneously.
-            /// Lists should be deserialized even when them overlaps
-            mod two_lists {
-                use super::*;
-
-                /// A field with a variable-name items defined before a field with a fixed-name
-                /// items
-                mod choice_and_fixed {
-                    use super::*;
-                    use pretty_assertions::assert_eq;
-
-                    #[derive(Debug, PartialEq, Deserialize)]
-                    struct Pair {
-                        #[serde(rename = "$value")]
-                        item: Vec<Choice>,
-                        element: Vec<()>,
-                    }
-
-                    /// A list with fixed-name elements located before a list with variable-name
-                    /// elements in an XML
-                    #[test]
-                    fn fixed_before() {
-                        let data: Pair = from_str(
-                            r#"
-                            <root>
-                                <element/>
-                                <element/>
-                                <one/>
-                                <two/>
-                                <three/>
-                            </root>
-                            "#,
-                        )
-                        .unwrap();
-
-                        assert_eq!(
-                            data,
-                            Pair {
-                                item: vec![Choice::One, Choice::Two, Choice::Other("three".into())],
-                                element: vec![(), ()],
-                            }
-                        );
-                    }
-
-                    /// A list with fixed-name elements located after a list with variable-name
-                    /// elements in an XML
-                    #[test]
-                    fn fixed_after() {
-                        let data: Pair = from_str(
-                            r#"
-                            <root>
-                                <one/>
-                                <two/>
-                                <three/>
-                                <element/>
-                                <element/>
-                            </root>
-                            "#,
-                        )
-                        .unwrap();
-
-                        assert_eq!(
-                            data,
-                            Pair {
-                                item: vec![Choice::One, Choice::Two, Choice::Other("three".into())],
-                                element: vec![(), ()],
-                            }
-                        );
-                    }
-
-                    mod overlapped {
-                        use super::*;
-                        use pretty_assertions::assert_eq;
-
-                        #[derive(Debug, PartialEq, Deserialize)]
-                        struct Root {
-                            #[serde(rename = "$value")]
-                            item: Vec<Choice4>,
-                            element: Vec<()>,
-                        }
-
-                        /// A list with fixed-name elements are mixed with a list with variable-name
-                        /// elements in an XML, and the first element is a fixed-name one
-                        #[test]
-                        fn fixed_before() {
-                            let data = from_str::<Pair>(
-                                r#"
-                                <root>
-                                    <element/>
-                                    <one/>
-                                    <two/>
-                                    <element/>
-                                    <three/>
-                                </root>
-                                "#,
-                            );
-
-                            #[cfg(feature = "overlapped-lists")]
-                            assert_eq!(
-                                data.unwrap(),
-                                Pair {
-                                    item: vec![
-                                        Choice::One,
-                                        Choice::Two,
-                                        Choice::Other("three".into())
-                                    ],
-                                    element: vec![(), ()],
-                                }
-                            );
-
-                            #[cfg(not(feature = "overlapped-lists"))]
-                            match data {
-                                Err(DeError::Custom(e)) => {
-                                    assert_eq!(e, "duplicate field `element`")
-                                }
-                                e => panic!(
-                                    r#"Expected Err(Custom("duplicate field `element`")), got {:?}"#,
-                                    e
-                                ),
-                            }
-                        }
-
-                        /// A list with fixed-name elements are mixed with a list with variable-name
-                        /// elements in an XML, and the first element is a variable-name one
-                        #[test]
-                        fn fixed_after() {
-                            let data = from_str::<Pair>(
-                                r#"
-                                <root>
-                                    <one/>
-                                    <element/>
-                                    <two/>
-                                    <three/>
-                                    <element/>
-                                </root>
-                                "#,
-                            );
-
-                            #[cfg(feature = "overlapped-lists")]
-                            assert_eq!(
-                                data.unwrap(),
-                                Pair {
-                                    item: vec![
-                                        Choice::One,
-                                        Choice::Two,
-                                        Choice::Other("three".into())
-                                    ],
-                                    element: vec![(), ()],
-                                }
-                            );
-
-                            #[cfg(not(feature = "overlapped-lists"))]
-                            match data {
-                                Err(DeError::Custom(e)) => {
-                                    assert_eq!(e, "duplicate field `$value`")
-                                }
-                                e => panic!(
-                                    r#"Expected Err(Custom("duplicate field `$value`")), got {:?}"#,
-                                    e
-                                ),
-                            }
-                        }
-
-                        /// Test for https://github.com/tafia/quick-xml/issues/435
-                        #[test]
-                        fn with_nested_list_fixed_before() {
-                            let data = from_str::<Root>(
-                                r#"
-                                <root>
-                                    <element/>
-                                    <one><inner/></one>
-                                    <two><inner/></two>
-                                    <element/>
-                                    <three><inner/></three>
-                                </root>
-                                "#,
-                            );
-
-                            #[cfg(feature = "overlapped-lists")]
-                            assert_eq!(
-                                data.unwrap(),
-                                Root {
-                                    item: vec![
-                                        Choice4::One { inner: [()] },
-                                        Choice4::Two { inner: [()] },
-                                        Choice4::Other,
-                                    ],
-                                    element: vec![(); 2],
-                                }
-                            );
-
-                            #[cfg(not(feature = "overlapped-lists"))]
-                            match data {
-                                Err(DeError::Custom(e)) => {
-                                    assert_eq!(e, "duplicate field `element`")
-                                }
-                                e => panic!(
-                                    r#"Expected Err(Custom("duplicate field `element`")), got {:?}"#,
-                                    e
-                                ),
-                            }
-                        }
-
-                        /// Test for https://github.com/tafia/quick-xml/issues/435
-                        #[test]
-                        fn with_nested_list_fixed_after() {
-                            let data = from_str::<Root>(
-                                r#"
-                                <root>
-                                    <one><inner/></one>
-                                    <element/>
-                                    <two><inner/></two>
-                                    <three><inner/></three>
-                                    <element/>
-                                </root>
-                                "#,
-                            );
-
-                            #[cfg(feature = "overlapped-lists")]
-                            assert_eq!(
-                                data.unwrap(),
-                                Root {
-                                    item: vec![
-                                        Choice4::One { inner: [()] },
-                                        Choice4::Two { inner: [()] },
-                                        Choice4::Other,
-                                    ],
-                                    element: vec![(); 2],
-                                }
-                            );
-
-                            #[cfg(not(feature = "overlapped-lists"))]
-                            match data {
-                                Err(DeError::Custom(e)) => {
-                                    assert_eq!(e, "duplicate field `$value`")
-                                }
-                                e => panic!(
-                                    r#"Expected Err(Custom("duplicate field `$value`")), got {:?}"#,
-                                    e
-                                ),
-                            }
-                        }
-                    }
-                }
-
-                /// A field with a variable-name items defined after a field with a fixed-name
-                /// items
-                mod fixed_and_choice {
-                    use super::*;
-                    use pretty_assertions::assert_eq;
-
-                    #[derive(Debug, PartialEq, Deserialize)]
-                    struct Pair {
-                        element: Vec<()>,
-                        #[serde(rename = "$value")]
-                        item: Vec<Choice>,
-                    }
-
-                    /// A list with fixed-name elements located before a list with variable-name
-                    /// elements in an XML
-                    #[test]
-                    fn fixed_before() {
-                        let data: Pair = from_str(
-                            r#"
-                            <root>
-                                <element/>
-                                <element/>
-                                <one/>
-                                <two/>
-                                <three/>
-                            </root>
-                            "#,
-                        )
-                        .unwrap();
-
-                        assert_eq!(
-                            data,
-                            Pair {
-                                element: vec![(), ()],
-                                item: vec![Choice::One, Choice::Two, Choice::Other("three".into())],
-                            }
-                        );
-                    }
-
-                    /// A list with fixed-name elements located after a list with variable-name
-                    /// elements in an XML
-                    #[test]
-                    fn fixed_after() {
-                        let data: Pair = from_str(
-                            r#"
-                            <root>
-                                <one/>
-                                <two/>
-                                <three/>
-                                <element/>
-                                <element/>
-                            </root>
-                            "#,
-                        )
-                        .unwrap();
-
-                        assert_eq!(
-                            data,
-                            Pair {
-                                element: vec![(), ()],
-                                item: vec![Choice::One, Choice::Two, Choice::Other("three".into())],
-                            }
-                        );
-                    }
-
-                    mod overlapped {
-                        use super::*;
-                        use pretty_assertions::assert_eq;
-
-                        #[derive(Debug, PartialEq, Deserialize)]
-                        struct Root {
-                            element: Vec<()>,
-                            #[serde(rename = "$value")]
-                            item: Vec<Choice4>,
-                        }
-
-                        /// A list with fixed-name elements are mixed with a list with variable-name
-                        /// elements in an XML, and the first element is a fixed-name one
-                        #[test]
-                        fn fixed_before() {
-                            let data = from_str::<Pair>(
-                                r#"
-                                <root>
-                                    <element/>
-                                    <one/>
-                                    <two/>
-                                    <element/>
-                                    <three/>
-                                </root>
-                                "#,
-                            );
-
-                            #[cfg(feature = "overlapped-lists")]
-                            assert_eq!(
-                                data.unwrap(),
-                                Pair {
-                                    element: vec![(), ()],
-                                    item: vec![
-                                        Choice::One,
-                                        Choice::Two,
-                                        Choice::Other("three".into())
-                                    ],
-                                }
-                            );
-
-                            #[cfg(not(feature = "overlapped-lists"))]
-                            match data {
-                                Err(DeError::Custom(e)) => {
-                                    assert_eq!(e, "duplicate field `element`")
-                                }
-                                e => panic!(
-                                    r#"Expected Err(Custom("duplicate field `element`")), got {:?}"#,
-                                    e
-                                ),
-                            }
-                        }
-
-                        /// A list with fixed-name elements are mixed with a list with variable-name
-                        /// elements in an XML, and the first element is a variable-name one
-                        #[test]
-                        fn fixed_after() {
-                            let data = from_str::<Pair>(
-                                r#"
-                                <root>
-                                    <one/>
-                                    <element/>
-                                    <two/>
-                                    <three/>
-                                    <element/>
-                                </root>
-                                "#,
-                            );
-
-                            #[cfg(feature = "overlapped-lists")]
-                            assert_eq!(
-                                data.unwrap(),
-                                Pair {
-                                    element: vec![(), ()],
-                                    item: vec![
-                                        Choice::One,
-                                        Choice::Two,
-                                        Choice::Other("three".into())
-                                    ],
-                                }
-                            );
-
-                            #[cfg(not(feature = "overlapped-lists"))]
-                            match data {
-                                Err(DeError::Custom(e)) => {
-                                    assert_eq!(e, "duplicate field `$value`")
-                                }
-                                e => panic!(
-                                    r#"Expected Err(Custom("duplicate field `$value`")), got {:?}"#,
-                                    e
-                                ),
-                            }
-                        }
-
-                        /// Test for https://github.com/tafia/quick-xml/issues/435
-                        #[test]
-                        fn with_nested_list_fixed_before() {
-                            let data = from_str::<Root>(
-                                r#"
-                                <root>
-                                    <element/>
-                                    <one><inner/></one>
-                                    <two><inner/></two>
-                                    <element/>
-                                    <three><inner/></three>
-                                </root>
-                                "#,
-                            );
-
-                            #[cfg(feature = "overlapped-lists")]
-                            assert_eq!(
-                                data.unwrap(),
-                                Root {
-                                    element: vec![(); 2],
-                                    item: vec![
-                                        Choice4::One { inner: [()] },
-                                        Choice4::Two { inner: [()] },
-                                        Choice4::Other,
-                                    ],
-                                }
-                            );
-
-                            #[cfg(not(feature = "overlapped-lists"))]
-                            match data {
-                                Err(DeError::Custom(e)) => {
-                                    assert_eq!(e, "duplicate field `element`")
-                                }
-                                e => panic!(
-                                    r#"Expected Err(Custom("duplicate field `element`")), got {:?}"#,
-                                    e
-                                ),
-                            }
-                        }
-
-                        /// Test for https://github.com/tafia/quick-xml/issues/435
-                        #[test]
-                        fn with_nested_list_fixed_after() {
-                            let data = from_str::<Root>(
-                                r#"
-                                <root>
-                                    <one><inner/></one>
-                                    <element/>
-                                    <two><inner/></two>
-                                    <three><inner/></three>
-                                    <element/>
-                                </root>
-                                "#,
-                            );
-
-                            #[cfg(feature = "overlapped-lists")]
-                            assert_eq!(
-                                data.unwrap(),
-                                Root {
-                                    element: vec![(); 2],
-                                    item: vec![
-                                        Choice4::One { inner: [()] },
-                                        Choice4::Two { inner: [()] },
-                                        Choice4::Other,
-                                    ],
-                                }
-                            );
-
-                            #[cfg(not(feature = "overlapped-lists"))]
-                            match data {
-                                Err(DeError::Custom(e)) => {
-                                    assert_eq!(e, "duplicate field `$value`")
-                                }
-                                e => panic!(
-                                    r#"Expected Err(Custom("duplicate field `$value`")), got {:?}"#,
-                                    e
-                                ),
-                            }
-                        }
-                    }
-                }
-
-                /// Tests are ignored, but exists to show a problem.
-                /// May be it will be solved in the future
-                mod choice_and_choice {
-                    use super::*;
-                    use pretty_assertions::assert_eq;
-
-                    #[derive(Debug, PartialEq, Deserialize)]
-                    struct Pair {
-                        #[serde(rename = "$value")]
-                        item: Vec<Choice>,
-                        // Actually, we cannot rename both fields to `$value`, which is now
-                        // required to indicate, that field accepts elements with any name
-                        #[serde(rename = "$value")]
-                        element: Vec<Choice2>,
-                    }
-
-                    #[test]
-                    #[ignore = "There is no way to associate XML elements with `item` or `element` without extra knowledge from type"]
-                    fn splitted() {
-                        let data: Pair = from_str(
-                            r#"
-                            <root>
-                                <first/>
-                                <second/>
-                                <one/>
-                                <two/>
-                                <three/>
-                            </root>
-                            "#,
-                        )
-                        .unwrap();
-
-                        assert_eq!(
-                            data,
-                            Pair {
-                                item: vec![Choice::One, Choice::Two, Choice::Other("three".into())],
-                                element: vec![Choice2::First, Choice2::Second],
-                            }
-                        );
-                    }
-
-                    #[test]
-                    #[ignore = "There is no way to associate XML elements with `item` or `element` without extra knowledge from type"]
-                    fn overlapped() {
-                        let data = from_str::<Pair>(
-                            r#"
-                            <root>
-                                <one/>
-                                <first/>
-                                <two/>
-                                <second/>
-                                <three/>
-                            </root>
-                            "#,
-                        );
-
-                        #[cfg(feature = "overlapped-lists")]
-                        assert_eq!(
-                            data.unwrap(),
-                            Pair {
-                                item: vec![Choice::One, Choice::Two, Choice::Other("three".into())],
-                                element: vec![Choice2::First, Choice2::Second],
-                            }
-                        );
-
-                        #[cfg(not(feature = "overlapped-lists"))]
-                        match data {
-                            Err(DeError::Custom(e)) => {
-                                assert_eq!(e, "invalid length 1, expected an array of length 3")
-                            }
-                            e => panic!(
-                                r#"Expected Err(Custom("invalid length 1, expected an array of length 3")), got {:?}"#,
-                                e
-                            ),
-                        }
-                    }
-                }
-            }
-
-            /// Deserialization of primitives slightly differs from deserialization
-            /// of complex types, so need to check this separately
-            #[test]
-            fn primitives() {
-                #[derive(Debug, PartialEq, Deserialize)]
-                struct List {
-                    #[serde(rename = "$value")]
-                    item: Vec<Choice3>,
-                }
-
-                let data: List = from_str(
-                    r#"
-                    <root>
-                        <one>41</one>
-                        <two>42</two>
-                        <three>43</three>
-                    </root>
-                    "#,
-                )
-                .unwrap();
-
-                assert_eq!(
-                    data,
-                    List {
-                        item: vec![
-                            Choice3::One(41),
-                            Choice3::Two("42".to_string()),
-                            Choice3::Other,
-                        ],
-                    }
-                );
-
-                from_str::<List>(
-                    r#"
-                    <root>
-                        <one>41</one>
-                        <two><item>42</item></two>
-                        <three>43</three>
-                    </root>
-                    "#,
-                )
-                .unwrap_err();
-            }
-
-            /// Checks that sequences represented by elements can contain sequences,
-            /// represented by `xs:list`s
-            mod xs_list {
-                use super::*;
-                use pretty_assertions::assert_eq;
-
-                #[derive(Debug, Deserialize, PartialEq)]
-                struct List {
-                    /// Outer list mapped to elements, inner -- to `xs:list`.
-                    ///
-                    /// `#[serde(default)]` is required to correctly deserialize
-                    /// empty sequence, because without elements the field
-                    /// also is missing and derived `Deserialize` implementation
-                    /// would complain about that unless field is marked as
-                    /// `default`.
-                    #[serde(default)]
-                    #[serde(rename = "$value")]
-                    element: Vec<Vec<String>>,
-                }
-
-                /// Special case: zero elements
-                #[test]
-                fn zero() {
-                    let data: List = from_str(
-                        r#"
-                        <root>
-                        </root>
-                        "#,
-                    )
-                    .unwrap();
-
-                    assert_eq!(data, List { element: vec![] });
-                }
-
-                /// Special case: one element
-                #[test]
-                fn one() {
-                    let data: List = from_str(
-                        r#"
-                        <root>
-                            <item>first list</item>
-                        </root>
-                        "#,
-                    )
-                    .unwrap();
-
-                    assert_eq!(
-                        data,
-                        List {
-                            element: vec![vec!["first".to_string(), "list".to_string()]]
-                        }
-                    );
-                }
-
-                /// Special case: outer list is always mapped to an elements sequence,
-                /// not to an `xs:list`
-                #[test]
-                fn element() {
-                    #[derive(Debug, Deserialize, PartialEq)]
-                    struct List {
-                        /// Outer list mapped to elements.
-                        #[serde(rename = "$value")]
-                        element: Vec<String>,
-                    }
-
-                    let data: List = from_str(
-                        r#"
-                        <root>
-                            <item>first item</item>
-                        </root>
-                        "#,
-                    )
-                    .unwrap();
-
-                    assert_eq!(
-                        data,
-                        List {
-                            element: vec!["first item".to_string()]
-                        }
-                    );
-                }
-
-                /// This tests demonstrates, that for `$value` field (`list`) actual
-                /// name of XML element (`item`) does not matter. That allows list
-                /// item to be an enum, where tag name determines enum variant
-                #[test]
-                fn many() {
-                    let data: List = from_str(
-                        r#"
-                        <root>
-                            <item>first list</item>
-                            <item>second list</item>
-                        </root>
-                        "#,
-                    )
-                    .unwrap();
-
-                    assert_eq!(
-                        data,
-                        List {
-                            element: vec![
-                                vec!["first".to_string(), "list".to_string()],
-                                vec!["second".to_string(), "list".to_string()],
-                            ]
-                        }
-                    );
-                }
-            }
-        }
-    }
-}
+// seq tests are so big, so it in the separate file serde-de-seq.rs to speed-up compilation
 
 macro_rules! maplike_errors {
     ($type:ty) => {
@@ -5058,1036 +961,7 @@ mod flatten_struct {
     }
 }
 
-mod enum_ {
-    use super::*;
-
-    #[derive(Debug, Deserialize, PartialEq)]
-    struct Nested {
-        //TODO: change to f64 after fixing https://github.com/serde-rs/serde/issues/1183
-        float: String,
-    }
-
-    /// Type where all struct fields represented by attributes
-    #[derive(Debug, Deserialize, PartialEq)]
-    struct NestedAttr {
-        //TODO: change to f64 after fixing https://github.com/serde-rs/serde/issues/1183
-        #[serde(rename = "@float")]
-        float: String,
-    }
-
-    /// Enum tag selector is a name of the `<element>` or text / CDATA content
-    /// for a `$text` variant
-    mod externally_tagged {
-        use super::*;
-        use pretty_assertions::assert_eq;
-
-        /// Type where all fields of struct variants represented by elements
-        #[derive(Debug, Deserialize, PartialEq)]
-        enum Node {
-            Unit,
-            Newtype(bool),
-            //TODO: serde bug https://github.com/serde-rs/serde/issues/1904
-            // Tuple(f64, String),
-            Struct {
-                float: f64,
-                string: String,
-            },
-            Holder {
-                nested: Nested,
-                string: String,
-            },
-            Flatten {
-                #[serde(flatten)]
-                nested: Nested,
-                string: String,
-            },
-        }
-
-        /// Type where all fields of struct variants represented by attributes
-        #[derive(Debug, Deserialize, PartialEq)]
-        enum NodeAttr {
-            Struct {
-                #[serde(rename = "@float")]
-                float: f64,
-                #[serde(rename = "@string")]
-                string: String,
-            },
-            Holder {
-                nested: NestedAttr,
-                #[serde(rename = "@string")]
-                string: String,
-            },
-            Flatten {
-                #[serde(flatten)]
-                nested: NestedAttr,
-                #[serde(rename = "@string")]
-                string: String,
-            },
-        }
-
-        /// Workaround for serde bug https://github.com/serde-rs/serde/issues/1904
-        #[derive(Debug, Deserialize, PartialEq)]
-        enum Workaround {
-            Tuple(f64, String),
-        }
-
-        #[test]
-        fn unit() {
-            let data: Node = from_str("<Unit/>").unwrap();
-            assert_eq!(data, Node::Unit);
-        }
-
-        #[test]
-        fn newtype() {
-            let data: Node = from_str("<Newtype>true</Newtype>").unwrap();
-            assert_eq!(data, Node::Newtype(true));
-        }
-
-        #[test]
-        fn tuple_struct() {
-            let data: Workaround = from_str("<Tuple>42</Tuple><Tuple>answer</Tuple>").unwrap();
-            assert_eq!(data, Workaround::Tuple(42.0, "answer".into()));
-        }
-
-        mod struct_ {
-            use super::*;
-            use pretty_assertions::assert_eq;
-
-            #[test]
-            fn elements() {
-                let data: Node = from_str(
-                    // Comment for prevent unnecessary formatting - we use the same style in all tests
-                    r#"<Struct><float>42</float><string>answer</string></Struct>"#,
-                )
-                .unwrap();
-                assert_eq!(
-                    data,
-                    Node::Struct {
-                        float: 42.0,
-                        string: "answer".into()
-                    }
-                );
-            }
-
-            #[test]
-            fn attributes() {
-                let data: NodeAttr = from_str(
-                    // Comment for prevent unnecessary formatting - we use the same style in all tests
-                    r#"<Struct float="42" string="answer"/>"#,
-                )
-                .unwrap();
-                assert_eq!(
-                    data,
-                    NodeAttr::Struct {
-                        float: 42.0,
-                        string: "answer".into()
-                    }
-                );
-            }
-
-            #[test]
-            fn namespaces() {
-                let data: Node = from_str(
-                    // Comment for prevent unnecessary formatting - we use the same style in all tests
-                    r#"<namespace:Struct xmlns:namespace="http://name.space"><float>42</float><string>answer</string></namespace:Struct>"#,
-                )
-                .unwrap();
-                assert_eq!(
-                    data,
-                    Node::Struct {
-                        float: 42.0,
-                        string: "answer".into()
-                    }
-                );
-            }
-        }
-
-        mod nested_struct {
-            use super::*;
-            use pretty_assertions::assert_eq;
-
-            #[test]
-            fn elements() {
-                let data: Node = from_str(
-                    r#"<Holder><string>answer</string><nested><float>42</float></nested></Holder>"#,
-                )
-                .unwrap();
-                assert_eq!(
-                    data,
-                    Node::Holder {
-                        nested: Nested { float: "42".into() },
-                        string: "answer".into()
-                    }
-                );
-            }
-
-            #[test]
-            fn attributes() {
-                let data: NodeAttr = from_str(
-                    // Comment for prevent unnecessary formatting - we use the same style in all tests
-                    r#"<Holder string="answer"><nested float="42"/></Holder>"#,
-                )
-                .unwrap();
-                assert_eq!(
-                    data,
-                    NodeAttr::Holder {
-                        nested: NestedAttr { float: "42".into() },
-                        string: "answer".into()
-                    }
-                );
-            }
-        }
-
-        mod flatten_struct {
-            use super::*;
-            use pretty_assertions::assert_eq;
-
-            #[test]
-            #[ignore = "Prime cause: deserialize_any under the hood + https://github.com/serde-rs/serde/issues/1183"]
-            fn elements() {
-                let data: Node = from_str(
-                    // Comment for prevent unnecessary formatting - we use the same style in all tests
-                    r#"<Flatten><float>42</float><string>answer</string></Flatten>"#,
-                )
-                .unwrap();
-                assert_eq!(
-                    data,
-                    Node::Flatten {
-                        nested: Nested { float: "42".into() },
-                        string: "answer".into()
-                    }
-                );
-            }
-
-            #[test]
-            fn attributes() {
-                let data: NodeAttr = from_str(
-                    // Comment for prevent unnecessary formatting - we use the same style in all tests
-                    r#"<Flatten float="42" string="answer"/>"#,
-                )
-                .unwrap();
-                assert_eq!(
-                    data,
-                    NodeAttr::Flatten {
-                        nested: NestedAttr { float: "42".into() },
-                        string: "answer".into()
-                    }
-                );
-            }
-        }
-
-        /// Test deserialization of the specially named variant `$text`
-        mod text {
-            use super::*;
-
-            mod unit {
-                use super::*;
-                use pretty_assertions::assert_eq;
-
-                #[derive(Debug, Deserialize, PartialEq)]
-                enum Text {
-                    #[serde(rename = "$text")]
-                    Unit,
-                }
-
-                #[test]
-                fn text() {
-                    let data: Text = from_str(" text ").unwrap();
-                    assert_eq!(data, Text::Unit);
-                }
-
-                #[test]
-                fn cdata() {
-                    let data: Text = from_str("<![CDATA[ cdata ]]>").unwrap();
-                    assert_eq!(data, Text::Unit);
-                }
-
-                #[test]
-                #[ignore = "awaiting fix of https://github.com/tafia/quick-xml/issues/474"]
-                fn mixed() {
-                    let data: Text = from_str(" te <![CDATA[ cdata ]]> xt ").unwrap();
-                    assert_eq!(data, Text::Unit);
-                }
-            }
-
-            mod newtype {
-                use super::*;
-                use pretty_assertions::assert_eq;
-
-                #[derive(Debug, Deserialize, PartialEq)]
-                enum Text {
-                    #[serde(rename = "$text")]
-                    Newtype(String),
-                }
-
-                #[test]
-                fn text() {
-                    let data: Text = from_str(" text ").unwrap();
-                    assert_eq!(data, Text::Newtype("text".into()));
-                }
-
-                #[test]
-                fn cdata() {
-                    let data: Text = from_str("<![CDATA[ cdata ]]>").unwrap();
-                    assert_eq!(data, Text::Newtype(" cdata ".into()));
-                }
-
-                #[test]
-                #[ignore = "awaiting fix of https://github.com/tafia/quick-xml/issues/474"]
-                fn mixed() {
-                    let data: Text = from_str(" te <![CDATA[ cdata ]]> xt ").unwrap();
-                    assert_eq!(data, Text::Newtype("te  cdata  xt".into()));
-                }
-            }
-
-            /// Tuple variant deserialized as an `xs:list`, that is why spaces
-            /// are trimmed even in CDATA sections
-            mod tuple {
-                use super::*;
-                use pretty_assertions::assert_eq;
-
-                #[derive(Debug, Deserialize, PartialEq)]
-                enum Text {
-                    #[serde(rename = "$text")]
-                    Tuple(f64, String),
-                }
-
-                #[test]
-                fn text() {
-                    let data: Text = from_str(" 4.2 text ").unwrap();
-                    assert_eq!(data, Text::Tuple(4.2, "text".into()));
-                }
-
-                #[test]
-                fn cdata() {
-                    let data: Text = from_str("<![CDATA[ 4.2 cdata ]]>").unwrap();
-                    assert_eq!(data, Text::Tuple(4.2, "cdata".into()));
-                }
-
-                #[test]
-                #[ignore = "awaiting fix of https://github.com/tafia/quick-xml/issues/474"]
-                fn mixed() {
-                    let data: Text = from_str(" 4.2 <![CDATA[ cdata ]]>").unwrap();
-                    assert_eq!(data, Text::Tuple(4.2, "cdata".into()));
-                }
-            }
-
-            /// Struct variant cannot be directly deserialized from `Text` / `CData` events
-            mod struct_ {
-                use super::*;
-
-                #[derive(Debug, Deserialize, PartialEq)]
-                enum Text {
-                    #[serde(rename = "$text")]
-                    Struct { float: f64, string: String },
-                }
-
-                #[test]
-                fn text() {
-                    match from_str::<Text>(" text ") {
-                        Err(DeError::Unsupported(_)) => {}
-                        x => panic!("Expected `Err(Unsupported(_))`, but found {:?}", x),
-                    }
-                }
-
-                #[test]
-                fn cdata() {
-                    match from_str::<Text>("<![CDATA[ cdata ]]>") {
-                        Err(DeError::Unsupported(_)) => {}
-                        x => panic!("Expected `Err(Unsupported(_))`, but found {:?}", x),
-                    }
-                }
-
-                #[test]
-                fn mixed() {
-                    match from_str::<Text>(" te <![CDATA[ cdata ]]> xt ") {
-                        Err(DeError::Unsupported(_)) => {}
-                        x => panic!("Expected `Err(Unsupported(_))`, but found {:?}", x),
-                    }
-                }
-            }
-        }
-    }
-
-    /// Enum tag selector either an attribute "tag", or a tag "tag".
-    /// `$text` variant could be defined, but that name has no special meaning
-    mod internally_tagged {
-        use super::*;
-
-        /// Type where all fields of struct variants and a tag represented by elements
-        #[derive(Debug, Deserialize, PartialEq)]
-        #[serde(tag = "tag")]
-        enum Node {
-            Unit,
-            /// Primitives (such as `bool`) are not supported by serde in the internally tagged mode
-            Newtype(NewtypeContent),
-            // Tuple(f64, String),// Tuples are not supported in the internally tagged mode
-            Struct {
-                //TODO: change to f64 after fixing https://github.com/serde-rs/serde/issues/1183
-                float: String,
-                string: String,
-            },
-            Holder {
-                nested: Nested,
-                string: String,
-            },
-            Flatten {
-                #[serde(flatten)]
-                nested: Nested,
-                string: String,
-            },
-        }
-
-        /// Type where all fields of struct variants and a tag represented by attributes
-        #[derive(Debug, Deserialize, PartialEq)]
-        #[serde(tag = "@tag")]
-        enum NodeAttr {
-            Unit,
-            /// Primitives (such as `bool`) are not supported by serde in the internally tagged mode
-            Newtype(NewtypeContent),
-            // Tuple(f64, String),// Tuples are not supported in the internally tagged mode
-            Struct {
-                //TODO: change to f64 after fixing https://github.com/serde-rs/serde/issues/1183
-                #[serde(rename = "@float")]
-                float: String,
-                #[serde(rename = "@string")]
-                string: String,
-            },
-            Holder {
-                nested: NestedAttr,
-                #[serde(rename = "@string")]
-                string: String,
-            },
-            Flatten {
-                #[serde(flatten)]
-                nested: NestedAttr,
-                #[serde(rename = "@string")]
-                string: String,
-            },
-        }
-
-        #[derive(Debug, Deserialize, PartialEq)]
-        struct NewtypeContent {
-            value: bool,
-        }
-
-        mod unit {
-            use super::*;
-            use pretty_assertions::assert_eq;
-
-            #[test]
-            fn elements() {
-                let data: Node = from_str(r#"<root><tag>Unit</tag></root>"#).unwrap();
-                assert_eq!(data, Node::Unit);
-            }
-
-            #[test]
-            fn attributes() {
-                let data: NodeAttr = from_str(r#"<root tag="Unit"/>"#).unwrap();
-                assert_eq!(data, NodeAttr::Unit);
-            }
-        }
-
-        mod newtype {
-            use super::*;
-            use pretty_assertions::assert_eq;
-
-            #[test]
-            #[ignore = "Prime cause: deserialize_any under the hood + https://github.com/serde-rs/serde/issues/1183"]
-            fn elements() {
-                let data: Node = from_str(
-                    // Comment for prevent unnecessary formatting - we use the same style in all tests
-                    r#"<root><tag>Newtype</tag><value>true</value></root>"#,
-                )
-                .unwrap();
-                assert_eq!(data, Node::Newtype(NewtypeContent { value: true }));
-            }
-
-            #[test]
-            #[ignore = "Prime cause: deserialize_any under the hood + https://github.com/serde-rs/serde/issues/1183"]
-            fn attributes() {
-                let data: NodeAttr = from_str(
-                    // Comment for prevent unnecessary formatting - we use the same style in all tests
-                    r#"<root tag="Newtype"><value>true</value></root>"#,
-                )
-                .unwrap();
-                assert_eq!(data, NodeAttr::Newtype(NewtypeContent { value: true }));
-            }
-        }
-
-        mod struct_ {
-            use super::*;
-            use pretty_assertions::assert_eq;
-
-            #[test]
-            #[ignore = "Prime cause: deserialize_any under the hood + https://github.com/serde-rs/serde/issues/1183"]
-            fn elements() {
-                let data: Node = from_str(
-                    r#"<root><tag>Struct</tag><float>42</float><string>answer</string></root>"#,
-                )
-                .unwrap();
-                assert_eq!(
-                    data,
-                    Node::Struct {
-                        float: "42".into(),
-                        string: "answer".into()
-                    }
-                );
-            }
-
-            #[test]
-            fn attributes() {
-                let data: NodeAttr = from_str(
-                    // Comment for prevent unnecessary formatting - we use the same style in all tests
-                    r#"<root tag="Struct" float="42" string="answer"/>"#,
-                )
-                .unwrap();
-                assert_eq!(
-                    data,
-                    NodeAttr::Struct {
-                        float: "42".into(),
-                        string: "answer".into()
-                    }
-                );
-            }
-        }
-
-        mod nested_struct {
-            use super::*;
-            use pretty_assertions::assert_eq;
-
-            #[test]
-            #[ignore = "Prime cause: deserialize_any under the hood + https://github.com/serde-rs/serde/issues/1183"]
-            fn elements() {
-                let data: Node = from_str(
-                    r#"<root><tag>Holder</tag><string>answer</string><nested><float>42</float></nested></root>"#,
-                ).unwrap();
-                assert_eq!(
-                    data,
-                    Node::Holder {
-                        nested: Nested { float: "42".into() },
-                        string: "answer".into()
-                    }
-                );
-            }
-
-            #[test]
-            fn attributes() {
-                let data: NodeAttr = from_str(
-                    // Comment for prevent unnecessary formatting - we use the same style in all tests
-                    r#"<root tag="Holder" string="answer"><nested float="42"/></root>"#,
-                )
-                .unwrap();
-                assert_eq!(
-                    data,
-                    NodeAttr::Holder {
-                        nested: NestedAttr { float: "42".into() },
-                        string: "answer".into()
-                    }
-                );
-            }
-        }
-
-        mod flatten_struct {
-            use super::*;
-            use pretty_assertions::assert_eq;
-
-            #[test]
-            #[ignore = "Prime cause: deserialize_any under the hood + https://github.com/serde-rs/serde/issues/1183"]
-            fn elements() {
-                let data: Node = from_str(
-                    r#"<root><tag>Flatten</tag><float>42</float><string>answer</string></root>"#,
-                )
-                .unwrap();
-                assert_eq!(
-                    data,
-                    Node::Flatten {
-                        nested: Nested { float: "42".into() },
-                        string: "answer".into()
-                    }
-                );
-            }
-
-            #[test]
-            fn attributes() {
-                let data: NodeAttr = from_str(
-                    // Comment for prevent unnecessary formatting - we use the same style in all tests
-                    r#"<root tag="Flatten" float="42" string="answer"/>"#,
-                )
-                .unwrap();
-                assert_eq!(
-                    data,
-                    NodeAttr::Flatten {
-                        nested: NestedAttr { float: "42".into() },
-                        string: "answer".into()
-                    }
-                );
-            }
-        }
-    }
-
-    /// Enum tag selector either an attribute "tag", or a tag "tag".
-    /// `$text` variant could be defined, but that name has no special meaning
-    mod adjacently_tagged {
-        use super::*;
-
-        /// Type where all fields of struct variants, tag and content fields
-        /// represented by elements
-        #[derive(Debug, Deserialize, PartialEq)]
-        #[serde(tag = "tag", content = "content")]
-        enum Node {
-            Unit,
-            Newtype(bool),
-            //TODO: serde bug https://github.com/serde-rs/serde/issues/1904
-            // Tuple(f64, String),
-            Struct {
-                float: f64,
-                string: String,
-            },
-            Holder {
-                nested: Nested,
-                string: String,
-            },
-            Flatten {
-                #[serde(flatten)]
-                nested: Nested,
-                string: String,
-            },
-        }
-
-        /// Type where all fields of struct variants, tag and content fields
-        /// represented by attributes
-        #[derive(Debug, Deserialize, PartialEq)]
-        #[serde(tag = "@tag", content = "@content")]
-        enum NodeAttrSimple {
-            Unit,
-            Newtype(bool),
-        }
-
-        /// Type where all fields of struct variants and a tag represented by attributes
-        /// content cannot be represented by attribute because this is a complex struct
-        #[derive(Debug, Deserialize, PartialEq)]
-        #[serde(tag = "@tag", content = "content")]
-        enum NodeAttrComplex {
-            //TODO: serde bug https://github.com/serde-rs/serde/issues/1904
-            // Tuple(f64, String),
-            Struct {
-                #[serde(rename = "@float")]
-                float: f64,
-                #[serde(rename = "@string")]
-                string: String,
-            },
-            Holder {
-                nested: NestedAttr,
-                #[serde(rename = "@string")]
-                string: String,
-            },
-            Flatten {
-                #[serde(flatten)]
-                nested: NestedAttr,
-                #[serde(rename = "@string")]
-                string: String,
-            },
-        }
-
-        /// Workaround for serde bug https://github.com/serde-rs/serde/issues/1904
-        #[derive(Debug, Deserialize, PartialEq)]
-        #[serde(tag = "tag", content = "content")]
-        enum Workaround {
-            Tuple(f64, String),
-        }
-
-        /// Workaround for serde bug https://github.com/serde-rs/serde/issues/1904
-        #[derive(Debug, Deserialize, PartialEq)]
-        #[serde(tag = "@tag", content = "@content")]
-        enum WorkaroundAttr {
-            Tuple(f64, String),
-        }
-
-        mod unit {
-            use super::*;
-            use pretty_assertions::assert_eq;
-
-            #[test]
-            fn elements() {
-                let data: Node = from_str(r#"<root><tag>Unit</tag></root>"#).unwrap();
-                assert_eq!(data, Node::Unit);
-            }
-
-            #[test]
-            fn attributes() {
-                let data: NodeAttrSimple = from_str(r#"<root tag="Unit"/>"#).unwrap();
-                assert_eq!(data, NodeAttrSimple::Unit);
-            }
-        }
-
-        mod newtype {
-            use super::*;
-            use pretty_assertions::assert_eq;
-
-            #[test]
-            fn elements() {
-                let data: Node = from_str(
-                    // Comment for prevent unnecessary formatting - we use the same style in all tests
-                    r#"<root><tag>Newtype</tag><content>true</content></root>"#,
-                )
-                .unwrap();
-                assert_eq!(data, Node::Newtype(true));
-            }
-
-            #[test]
-            fn attributes() {
-                let data: NodeAttrSimple = from_str(
-                    // Comment for prevent unnecessary formatting - we use the same style in all tests
-                    r#"<root tag="Newtype" content="true"/>"#,
-                )
-                .unwrap();
-                assert_eq!(data, NodeAttrSimple::Newtype(true));
-            }
-        }
-
-        mod tuple_struct {
-            use super::*;
-            use pretty_assertions::assert_eq;
-
-            #[test]
-            fn elements() {
-                let data: Workaround = from_str(
-                    r#"<root><tag>Tuple</tag><content>42</content><content>answer</content></root>"#,
-                ).unwrap();
-                assert_eq!(data, Workaround::Tuple(42.0, "answer".into()));
-            }
-
-            #[test]
-            #[ignore = "Prime cause: deserialize_any under the hood + https://github.com/serde-rs/serde/issues/1183"]
-            fn attributes() {
-                let data: WorkaroundAttr = from_str(
-                    // We cannot have two attributes with the same name, so both values stored in one attribute
-                    r#"<root tag="Tuple" content="42 answer"/>"#,
-                )
-                .unwrap();
-                assert_eq!(data, WorkaroundAttr::Tuple(42.0, "answer".into()));
-            }
-        }
-
-        mod struct_ {
-            use super::*;
-            use pretty_assertions::assert_eq;
-
-            #[test]
-            fn elements() {
-                let data: Node = from_str(
-                    r#"<root><tag>Struct</tag><content><float>42</float><string>answer</string></content></root>"#,
-                )
-                .unwrap();
-                assert_eq!(
-                    data,
-                    Node::Struct {
-                        float: 42.0,
-                        string: "answer".into()
-                    }
-                );
-            }
-
-            #[test]
-            fn attributes() {
-                let data: NodeAttrComplex = from_str(
-                    // Comment for prevent unnecessary formatting - we use the same style in all tests
-                    r#"<root tag="Struct"><content float="42" string="answer"/></root>"#,
-                )
-                .unwrap();
-                assert_eq!(
-                    data,
-                    NodeAttrComplex::Struct {
-                        float: 42.0,
-                        string: "answer".into()
-                    }
-                );
-            }
-        }
-
-        mod nested_struct {
-            use super::*;
-            use pretty_assertions::assert_eq;
-
-            #[test]
-            fn elements() {
-                let data: Node = from_str(
-                    r#"<root>
-                        <tag>Holder</tag>
-                        <content>
-                            <string>answer</string>
-                            <nested>
-                                <float>42</float>
-                            </nested>
-                        </content>
-                    </root>"#,
-                )
-                .unwrap();
-                assert_eq!(
-                    data,
-                    Node::Holder {
-                        nested: Nested { float: "42".into() },
-                        string: "answer".into()
-                    }
-                );
-            }
-
-            #[test]
-            fn attributes() {
-                let data: NodeAttrComplex = from_str(
-                    r#"<root tag="Holder"><content string="answer"><nested float="42"/></content></root>"#,
-                ).unwrap();
-                assert_eq!(
-                    data,
-                    NodeAttrComplex::Holder {
-                        nested: NestedAttr { float: "42".into() },
-                        string: "answer".into()
-                    }
-                );
-            }
-        }
-
-        mod flatten_struct {
-            use super::*;
-            use pretty_assertions::assert_eq;
-
-            #[test]
-            #[ignore = "Prime cause: deserialize_any under the hood + https://github.com/serde-rs/serde/issues/1183"]
-            fn elements() {
-                let data: Node = from_str(
-                    r#"<root><tag>Flatten</tag><content><float>42</float><string>answer</string></content></root>"#,
-                ).unwrap();
-                assert_eq!(
-                    data,
-                    Node::Flatten {
-                        nested: Nested { float: "42".into() },
-                        string: "answer".into()
-                    }
-                );
-            }
-
-            #[test]
-            fn attributes() {
-                let data: NodeAttrComplex = from_str(
-                    // Comment for prevent unnecessary formatting - we use the same style in all tests
-                    r#"<root tag="Flatten"><content float="42" string="answer"/></root>"#,
-                )
-                .unwrap();
-                assert_eq!(
-                    data,
-                    NodeAttrComplex::Flatten {
-                        nested: NestedAttr { float: "42".into() },
-                        string: "answer".into()
-                    }
-                );
-            }
-        }
-    }
-
-    /// Enum tags does not exist.
-    /// `$text` variant could be defined, but that name has no special meaning
-    mod untagged {
-        use super::*;
-        use pretty_assertions::assert_eq;
-
-        /// Type where all fields of struct variants represented by elements
-        #[derive(Debug, Deserialize, PartialEq)]
-        #[serde(untagged)]
-        enum Node {
-            Unit,
-            Newtype(bool),
-            // serde bug https://github.com/serde-rs/serde/issues/1904
-            // Tuple(f64, String),
-            Struct {
-                float: f64,
-                string: String,
-            },
-            Holder {
-                nested: Nested,
-                string: String,
-            },
-            Flatten {
-                #[serde(flatten)]
-                nested: Nested,
-                // Can't use "string" as name because in that case this variant
-                // will have no difference from `Struct` variant
-                string2: String,
-            },
-        }
-
-        /// Type where all fields of struct variants represented by attributes
-        #[derive(Debug, Deserialize, PartialEq)]
-        #[serde(untagged)]
-        enum NodeAttr {
-            // serde bug https://github.com/serde-rs/serde/issues/1904
-            // Tuple(f64, String),
-            Struct {
-                #[serde(rename = "@float")]
-                float: f64,
-                #[serde(rename = "@string")]
-                string: String,
-            },
-            Holder {
-                nested: NestedAttr,
-                #[serde(rename = "@string")]
-                string: String,
-            },
-            Flatten {
-                #[serde(flatten)]
-                nested: NestedAttr,
-                // Can't use "string" as name because in that case this variant
-                // will have no difference from `Struct` variant
-                #[serde(rename = "@string2")]
-                string2: String,
-            },
-        }
-
-        /// Workaround for serde bug https://github.com/serde-rs/serde/issues/1904
-        #[derive(Debug, Deserialize, PartialEq)]
-        #[serde(untagged)]
-        enum Workaround {
-            Tuple(f64, String),
-        }
-
-        #[test]
-        #[ignore = "Prime cause: deserialize_any under the hood + https://github.com/serde-rs/serde/issues/1183"]
-        fn unit() {
-            // Unit variant consists just from the tag, and because tags
-            // are not written, nothing is written
-            let data: Node = from_str("").unwrap();
-            assert_eq!(data, Node::Unit);
-        }
-
-        #[test]
-        #[ignore = "Prime cause: deserialize_any under the hood + https://github.com/serde-rs/serde/issues/1183"]
-        fn newtype() {
-            let data: Node = from_str("true").unwrap();
-            assert_eq!(data, Node::Newtype(true));
-        }
-
-        #[test]
-        #[ignore = "Prime cause: deserialize_any under the hood + https://github.com/serde-rs/serde/issues/1183"]
-        fn tuple_struct() {
-            let data: Workaround = from_str("<root>42</root><root>answer</root>").unwrap();
-            assert_eq!(data, Workaround::Tuple(42.0, "answer".into()));
-        }
-
-        mod struct_ {
-            use super::*;
-            use pretty_assertions::assert_eq;
-
-            #[test]
-            #[ignore = "Prime cause: deserialize_any under the hood + https://github.com/serde-rs/serde/issues/1183"]
-            fn elements() {
-                let data: Node = from_str(
-                    // Comment for prevent unnecessary formatting - we use the same style in all tests
-                    r#"<root><float>42</float><string>answer</string></root>"#,
-                )
-                .unwrap();
-                assert_eq!(
-                    data,
-                    Node::Struct {
-                        float: 42.0,
-                        string: "answer".into()
-                    }
-                );
-            }
-
-            #[test]
-            #[ignore = "Prime cause: deserialize_any under the hood + https://github.com/serde-rs/serde/issues/1183"]
-            fn attributes() {
-                let data: NodeAttr = from_str(
-                    // Comment for prevent unnecessary formatting - we use the same style in all tests
-                    r#"<root float="42" string="answer"/>"#,
-                )
-                .unwrap();
-                assert_eq!(
-                    data,
-                    NodeAttr::Struct {
-                        float: 42.0,
-                        string: "answer".into()
-                    }
-                );
-            }
-        }
-
-        mod nested_struct {
-            use super::*;
-            use pretty_assertions::assert_eq;
-
-            #[test]
-            #[ignore = "Prime cause: deserialize_any under the hood + https://github.com/serde-rs/serde/issues/1183"]
-            fn elements() {
-                let data: Node = from_str(
-                    r#"<root><string>answer</string><nested><float>42</float></nested></root>"#,
-                )
-                .unwrap();
-                assert_eq!(
-                    data,
-                    Node::Holder {
-                        nested: Nested { float: "42".into() },
-                        string: "answer".into()
-                    }
-                );
-            }
-
-            #[test]
-            fn attributes() {
-                let data: NodeAttr = from_str(
-                    // Comment for prevent unnecessary formatting - we use the same style in all tests
-                    r#"<root string="answer"><nested float="42"/></root>"#,
-                )
-                .unwrap();
-                assert_eq!(
-                    data,
-                    NodeAttr::Holder {
-                        nested: NestedAttr { float: "42".into() },
-                        string: "answer".into()
-                    }
-                );
-            }
-        }
-
-        mod flatten_struct {
-            use super::*;
-            use pretty_assertions::assert_eq;
-
-            #[test]
-            #[ignore = "Prime cause: deserialize_any under the hood + https://github.com/serde-rs/serde/issues/1183"]
-            fn elements() {
-                let data: Node = from_str(
-                    // Comment for prevent unnecessary formatting - we use the same style in all tests
-                    r#"<root><float>42</float><string2>answer</string2></root>"#,
-                )
-                .unwrap();
-                assert_eq!(
-                    data,
-                    Node::Flatten {
-                        nested: Nested { float: "42".into() },
-                        string2: "answer".into()
-                    }
-                );
-            }
-
-            #[test]
-            fn attributes() {
-                let data: NodeAttr = from_str(
-                    // Comment for prevent unnecessary formatting - we use the same style in all tests
-                    r#"<root float="42" string2="answer"/>"#,
-                )
-                .unwrap();
-                assert_eq!(
-                    data,
-                    NodeAttr::Flatten {
-                        nested: NestedAttr { float: "42".into() },
-                        string2: "answer".into()
-                    }
-                );
-            }
-        }
-    }
-}
+// enum tests are so big, so it in the separate file serde-de-seq.rs to speed-up compilation
 
 /// https://www.w3schools.com/xml/el_list.asp
 mod xml_schema_lists {
